zephyr: zmake: use -j1 until we switch to GNU Make
Ninja doesn't know how to talk to a jobserver, and so all of our
ninjas go out of control, each trying to hit the CPU on all available
cores.  We have enough builds now that this is actually becoming
problematic as well.

-j1 forces sequential execution of the ninjas, which will mitigate the
problem until we switch to GNU Make.

BUG=b:178196029
BRANCH=none
TEST="zmake -D testall" executes sequentially

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I45516e22cca150a99f2da7a7b784d15c3fc78aab
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3075401
Reviewed-by: Simon Glass <sjg@chromium.org>