ps8xxx: Refactor role control delay logic
This refactors how we deal with ps8xxx_role_control_delay. An array of
booelans is replaced with an array of delay values with default of 0ms.

Instead of checking the array to determine whether a 1ms delay should be
performed, the array unconditionally specifes the delay to be performed.
The default 0ms delay is equivalent to the previous "false" entry.

BRANCH=none
BUG=b:186189039
TEST=buildall passes

Change-Id: Ie351c58442eb206f38cea7a1f30dd1789a8c8025
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3084330
Reviewed-by: Boris Mittelberg <bmbm@google.com>