ps8xxx: Wait for chip to wake up from LPM
The ps8815 can take up to 50ms to fully wake up from sleep/low power
mode. When the chip is asleep, the 1st I2C transaction will fail but the
chip will begin to wake up within 10ms. After this delay, I2C
transactions succeed, but the firmware is still not fully
operational. The way to check if the firmware is ready, is to poll the
firmware register for a non-zero value.

BRANCH=none
BUG=b:195087071,b:186189039
TEST=buildall passes

Change-Id: If047fc122d7f61ed5fc361f97b47180e5cf08970
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3084331
Reviewed-by: Keith Short <keithshort@chromium.org>