ANX7451: Report failure to set mux in G3
Since the ANX7451 is unpowered in Z1/G3, return an error for mux sets
during this time.  This will let the usb_mux.c code know that any set it
attempted to perform was unsuccessful.  Otherwise, it will be assumed
that setting mux mode of None succeeded, when we'll actually power up to
default USB mode.

BRANCH=None
BUG=b:195045790
TEST=on guybrush, power on board and ensure that C1 is put into None
mode successfully after its powered

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ic98be3bc15e8f0affeda25d7d7170b89e3aecf10
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3093087
Reviewed-by: Rob Barnes <robbarnes@google.com>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>