servo_v4p1: Initial support for FASTBOOT handling
Hardware support for FASTBOOT is in place, however servo_v4p1 is
currently not exposing an API for servod. Along with improvement for
correct DATA_ROLE handling on servo_v4p1 (in consecutive patch), add
a knob for enabling FASTBOOT.

It is necessary to add complete sequence when enabling FASTBOOT - that
is changing servo DATA_ROLE to DFP. This work will be done in the
future.

BUG=b:137887386
BRANCH=main
TEST=Conduct test from consecutive patch

Signed-off-by: Jan Dabros <jsd@semihalf.com>
Change-Id: I6cb53ebc9b412002db750b58d04170060478ebd4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3094248
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>