spherion: enable USM
Ultrasonic Module(USM) is used for noise cancelling the buzzing
sound while charging. The sound is louder if we have
USB peripheral connected. We only enable this function in S0,
and disable it in S3 because the power consumption is significant.

BUG=b:185978753
TEST=make buildall
BRANCH=asurada

Change-Id: Ic2bfbf5a1364f5da6d776a1be56abb66934f9496
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3105905
Tested-by: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@google.com>