Reland "zephyr: pull in the version string from zmake"
This reverts commit 93adb802c1a626fbb73d762124e90be4fee08a8b.

This commit was reverted due to build failures with the
chromeos-zephyr ebuild.  The root cause of the build failures was that
the ebuild uses different working directories between the configure
and build phases, and so the path to the include directory we generate
during the configure phase won't work in the build phase.

Use the full path here so that we can run the diffrerent phases in
different working directories.

Diff from original commit:

Change-Id: Idcbd0c0546f9fcb12fcd66c617e024b8aa2a893a

> --- a/zephyr/zmake/zmake/zmake.py
> +++ b/zephyr/zmake/zmake/zmake.py
> @@ -235,7 +235,7 @@ class Zmake:
>              self.logger.info("Clearing old build directory %s", build_dir)
>              shutil.rmtree(build_dir)
>
> -        generated_include_dir = build_dir / "include"
> +        generated_include_dir = (build_dir / "include").resolve()
>          base_config = zmake.build_config.BuildConfig(
>              environ_defs={"ZEPHYR_BASE": str(zephyr_base), "PATH": "/usr/bin"},
>              cmake_defs={

BUG=b:197287679
BRANCH=none
TEST=emerge-volteer chromeos-zephyr

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I80cec7a1622edcecc670b7dbeed56dadcdf71555
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3111305
Commit-Queue: Yuval Peress <peress@chromium.org>
Reviewed-by: Yuval Peress <peress@chromium.org>