zephyr: herobrine_npcx9: Enable PSL and define psl-in-pads binding
Enable the PSL mode for EC hibernate and define psl-in-pads DT binding.
After EC enters hibernate, the PSL_OUT is off and the PPC chip is
powered off. So the PPC chip can be woken up from the dead battery
mode when the external AC is plugged.

The DT binding is defined in the project DT. It enables the board DT
more shareable, no custom-board configuration.

BRANCH=None
BUG=b:193583152, b:196405396
TEST=Entered EC hibernate, verified the following wake sources:
 * plugging AC,
 * pressing power button,
 * lid open (servo emulated).

Change-Id: Ibe12d6ec45980cd41a306ad01f39e45456f3bcdd
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3115437
Reviewed-by: Keith Short <keithshort@chromium.org>