brask: add recovery button support
There is a dedicated recovery button for the brask.
This patch adds the CONFIG_DEDICATED_RECOVERY_BUTTON and
CONFIG_DEDICATED_RECOVERY_BUTTON_2 to enable it.
Besides, the CONFIG_EMULATED_SYSRQ and CONFIG_MKBP_INPUT_DEVICES
are added to let the EC pass the button state to the AP

BUG=b:191637086
BRANCH=None
TEST=make BOARD=brask

Change-Id: Ic6ea693bbbc0bcd0d502d2636636bc38e080b1e9
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3114500
Reviewed-by: Boris Mittelberg <bmbm@google.com>