retry charger init when on dead battery condition
The EC can power up before the charger IC has init when powered from
AC on disconnected battery. So we need to retry the charger init in this case.
Also track the PD PCH_PWR_EN to make sure charger has powered up before
we try to enable PCH_PWR, as it needs power output from the charger to
be present before this is allowed

Signed-off-by: Kieran Levin <ktl@frame.work>