Zephyr: Implement a generic battery led SM
Implement a generic battery led state machine for gpio led using device
tree. Currently this is implemented for lazor device as a first step.
It would be eventually converted to a standalone driver. This will be
applicable across all boards removing the need to have led.c
implementation for each board.

BRANCH=none
BUG=b:201007290
TEST=compared logs to verify the color value in each state

Signed-off-by: Parth Malkan <parthmalkan@google.com>
Change-Id: I571937457f9eb41d5949ffb8d1382d89ee118ba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3180480
Reviewed-by: Wai-Hong Tam <waihong@google.com>