zephyr: shim: handle sidesel in the shim file
This adds handling for the sidesel channel from the shim driver, if the
sidesel node is specified.

With this, all the pwm-led combinations should be manageable from the
shim driver, so this also removes the conditionals from volteer driver
and drops it from the Zephyr build entirely.

BRANCH=none
BUG=b:177452529
TEST=build and run on volteer
TEST=checked the binary output if node not defined

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Id215d26a8b14330b9604c7385d6ce436ef1a60cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3207995
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>