Revert "Clear the wake source flag to avoid the wrong pwoer state"
This reverts commit 4d241265a32777c86a325f9da402d59679365620.

EC code doesn't return the POWER_G3 state, so when we forcing_shutdown
and the flag is cleared early, it need to wait 10s down to S5G3 state.
If users press the power button, system won't power on and LED turn on,
it will cause the users confuse.

So we need to seprate the custom_forcing from common forcing_shutdown,
to keep pch power and avoid impact by common forcing_shutdwon.

Signed-off-by: Josh Tsai <josh_tsai@compal.corp-partner.google.com>