zephyr: test: usb_mux: Disable PD task during testing
A bug that was uncovered during crrev/c/3197935/8 where an extra sleep
was needed in an unrelated test would cause usb_mux test to fail if
removed.

ROOT CAUSE:
The PD task would make a call to (struct usb_mux_driver)::set as a
result to some stimulous. Normally during the test's run, this didn't
coincide with the usb_mux test suite so it didn't matter. The addition
of another test forced the threading overlap. When set() was called,
the assert would fail since the specific test wasn't expecting the call.

FIX:
Add logic to task.h to get the Zephyr TID from a given cros TASK_ID_*.
This function is used in the ztest setup/teardown methods of the
usb_mux test suite to turn on/off the PD tasks so that this race
condition is avoided.

BRANCH=none
BUG=none
TEST=zmake configure --test zephyr/test/drivers

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: I25adf0663fb89cd52a8677192e720932f0253626
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3213616
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>
Reviewed-by: Tomasz Michalec <tmichalec@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>