zmake: fail gracefully if missing SDK
The documentation didn't comment on the need to install a Zephyr SDK
package when working outside the chroot, and didn't provide any useful
output if one was missing (just failing with an assertion). Add mention
of the SDK to the documentation and fail with a more useful message if
it's missing and requested.

BUG=None
TEST=Useful message is shown if SDK is missing
BRANCH=None

Change-Id: I05fee5294c27197d62b3f726cdd5579571ce9e1b
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3216964
Reviewed-by: Andrew McRae <amcrae@google.com>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>