core/cortex-m: Fix RW image alignment when using clang (lld)
This change forces the LMA and VMA to start immediately after the .data
section; both LMA and VMA are forced to be equal. The change removes the
BYTE(0xFF), since lld will allocate space for this byte and throw off
the alignment.

The ASSERT line verifies that the __image_size is still correctly
aligned when building with both gcc and clang.

BRANCH=none
BUG=b:172020503
TEST=make buildall
TEST=make CC=arm-none-eabi-clang BOARD=dartmonkey

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I00d4e8f346232d1c5b8b04d397459760976a8a8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3169965
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@google.com>