tree: Remove .bss.slow
The config option to enable .bss.slow was removed in commit
b224441aaaa5dd08274811d65bb795b4abc231d8 because the option was only
used with kunimitsu, which was removed from the EC codebase a long time
ago.

Remove the reference in the linker maps to complete the cleanup.

BRANCH=none
BUG=b:172020503, b:172221339
TEST=make buildall

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Ibc519f9974cb26e6d15180e2ce0c359867ced1ac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3219263
Reviewed-by: Ting Shen <phoenixshen@chromium.org>