cbi-util: Fix error with setting 8 byte cbi data
From the ec console, setting 8 byte size rework id value using
'cbi set <tag> value size' fails for the tag CBI_TAG_REWORK_ID,
which is of 8byte size while rest of the structure members are
of 4 byte size or variable length.

This patch is to fix the error and allow setting of 8 byte cbi
value for the rework_id.

BUG=b:202870018
BRANCH=None
TEST=Run the command 'cbi set 9 0x123456789abcd 8' from
ec console and see it went through without errors.

Run the command 'cbi' which dumps all the fields of cbi
structure to check if the reworks_id(with tag 9) is
updated to this set value.

Signed-off-by: Thejaswani Putta <thejaswani.putta@intel.com>
Change-Id: Ifaa9c19dd5ea692b51c2666cdd37da09d395cf34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3219758
Reviewed-by: Thejaswani Putta <thejaswani.putta@intel.corp-partner.google.com>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Vijay Hiremath <vijay.p.hiremath@intel.com>
Tested-by: Thejaswani Putta <thejaswani.putta@intel.corp-partner.google.com>
Commit-Queue: Daisuke Nojiri <dnojiri@chromium.org>