set battery present only if valid SOC
Fix issue where the system would not power on if the battery was below
the low safety cutoff threshold, as we would detect the battery and mark
it as present, however the battery cutoff fet would still keep the
battery disabled as it was in precharge.

This would prevent the system from powering on.

Signed-off-by: Kieran Levin <ktl@frame.work>