zephyr: Test accelgyro_bmi260.c for unsupported operations
Add a test that covers parts of the code where we validate sensor types.

Also made a modification to accelgyro_bmi260.c where I moved the sensor
type check up to the top of the func so we fail before trying to set
the data rate (which doesn't work for magnetometers).

BRANCH=None
BUG=b:184856157
TEST=zmake configure --test zephyr/test/drivers

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I170087b8bfdc1212972d00010c90789967a6415b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3229076
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>