zephyr: Remove redundant check in accelgyro_bmi260.c:config_interrupt
Both `config_interrupt` and its sole caller check that the sensor struct
is for an accelerometer. This is redundant and creates an unreachable
code path that we cannot unit test. Remove the extra check and add
a comment to the function clarifying that it is the caller's
responsibility to verify the sensor is an accelerometer.

BRANCH=None
BUG=b:184856157
TEST=zmake configure --test zephyr/test/drivers

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I7c90fa11a1bf1ad9fbf52c88291115cceaf06687
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3232755
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>