driver: bmi3xx: Fix base accelerometer streaming issue
The configuration of sensor has to be saved before writing
offsets and it has to be written back.
Instead of writing back the saved configurations,
there was a call to read back the configs again.
This patch has the correct write API call.

BRANCH=none
BUG=b:200776764
TEST=1. make BOARD=guybrush  -j
     2. Flash EC binary on the gurbrush proto 2 device
     3. In kernel run:
         ectool motionsense
         cd /sys/bus/iio/devices/iio\:device2 && cat location
         echo 0 > in_accel_x_calibbias
         ectool motionsense # this should work

Signed-off-by: Latchiamaran Senram <latchiamaran.senram@bosch.corp-partner.google.com>

Change-Id: If8a20f50df8287fb49dec242f6ee9030e8bbe5d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3255475
Tested-by: Chih-Yao Chuang <jasonchuang@google.com>
Reviewed-by: Bhanu Prakash Maiya <bhanumaiya@google.com>
Commit-Queue: Bhanu Prakash Maiya <bhanumaiya@google.com>