TCPCI: Add an interface for I2C wake
When we wake a TCPC from I2C idle, there is a procedure in the TCPCI
specificaiton to follow.  Implementing the full procedure will take some
refactoring, but for now at least send the i2c wake to the chip when we
plan to wake it.  Otherwise, we don't actually wake the chip until the
LPM exit debounce finishes, which isn't in the intended use of the LPM
exit debounce.

BRANCH=None
BUG=b:195393479
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I6d5a6d27cce653ab02b59ab0e09d06657f640c39
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3262585
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>