Zephyr: Create sysworkq entry in task shim
At the moment, the sysworkq is sharing a task ID with the HOOKS task.
However, it needs its own entry in the dynamic task information in order
to receive events from other tasks. This entry doesn't need a thread
created, but does need an accurate thread ID and a polling entry.

BRANCH=None
BUG=b:195137794
TEST=on gubrush, verify sysworkq can now be notified of TCPC wakes and
charging works

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I89458be79e95f5f3c03ca67609b457c68e249b17
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3263321
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>