bq25710: Add VSYS_MIN to register definitions
This adds VSYS_MIN register definitions to bq257x0_regs.h. This register
is called MIN_SYSTEM_VOLTAGE on the bq25710, so that has also been
added for completeness.

Code referring to this register has been updated to use the chip
specific macros as the layout and encoding of the voltage field is
different.

BRANCH=none
BUG=b:185190976
TEST=buildall passes

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Change-Id: Ib456d66e10cb274282deca11a15262e8242ddb08
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3284639
Reviewed-by: Keith Short <keithshort@chromium.org>