usb_mux: tusb1064: Add method to set DP Rx EQ gain
This CL adds a method to set DP Rx EQ gain. The previous version of
the driver was enabling a 10 dB gain by default in the init
function. This CL removes setting both gain and EQ_OVERRIDE by
default.

A separate CL in the stack moves the 10dB gain setting to the board.c
file for the only other board that currently is using the TUSB1064.

BUG=b:206059703
BRANCH=quiche
TEST=Verified on gingerbread that both EQ_OVERRIDE is set and the
gain is set in both registers as expected.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Id2dc6127d27304d8149f7c51a8d527e046595baf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3291129
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Scott Collyer <scollyer@chromium.org>