zephyr: zmake: Drop support for per-project zephyr versioning
This drops all support for projects to specify which zephyr version
they use.  Zmake no longer cares what Zephyr version you give it: it
will simply use whatever's passed for --zephyr-base and be happy with
it.

This intentionally hard-codes the checkout path to the v2.7 branch as
the Zephyr version used to de-couple from the cutover to the main
branch.  The follow up CL actually switches to the main branch.  (This
is, more or less, just self-contained preparatory work in zmake for the
transition to main.)

BUG=b:205884929
BRANCH=none
TEST=unit tests pass
TEST=https://gitlab.com/zephyr-ec/ec/-/pipelines/412252730

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I9f8e28cec6466c666862c8f46f1608625a3ec53f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3296530
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>