core/cortex-m/task.c: Handle generalized EXC_RETURN
Previous version of the Cortex-M architecture enumerated a few
possible values of the lr register, upon entering an exception
handler.  The value would indicate aspects of the processor state
before being interrupted, and correspondingly how the core should
restore state when that value was loaded into the pc register upon
completion of the exception handler.

The values listed for M0 were:
0xFFFFFFF1: Return to Handler mode.
0xFFFFFFF9: Return to Thread mode.
0xFFFFFFFD: Return to Thread mode.
All other values: Reserved.

(There are differences between the two "Thread mode" value, which are
not important for this discussion.)

However, while developing code for STM32L5xx, which is Cortex-M33, I
encountered other values besides the above, and it turns out bit 3
indicates whether to return to thread mode or handler mode.  See
section 2.4.7.2 in this document for details:
https://www.st.com/resource/en/programming_manual/pm0264-stm32-cortexm33-mcus-programming-manual-stmicroelectronics.pdf

With the above knowledge, I have updated the condition in task.c to
handle previously reserved values.

BRANCH=none
BUG=b:192262089
TEST=Stress test on STM32L552 Nucleo board

Change-Id: If9b1995dad39cc87490bd825ee7e35a096712923
Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3297049
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>