zephyr: use a chosen node to detect internal flash
This change defines a well-known chosen node to bind the crec_flash_*
shim functions to an actual flash device. Prior to this change, the shim
required the internal flash be named 'fiu0' in order to function and
this would only be checked at runtime. It is now checked at
compile-time, saving a pointer in RAM and failing at build-time if not
correctly configured.

BUG=b:205615358
TEST=zmake testall
BRANCH=none

Change-Id: Ie713176960150a7ebeae8d2dcfc10c48260b264c
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3305647
Commit-Queue: Keith Short <keithshort@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>