zephyr: use a chosen device for raw keyboard shim
Rather than assuming the keyboard is named cros_kb_raw, get the device
from a chosen cros-ec,raw-kb node instead. To make this work whether
the keyboard is in use or not, also move the cros_kb_raw node to live in
the main it8xxx2.dtsi file, rather than being duplicated for those
boards that use it.

BUG=b:205615358
TEST=zmake testall
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Id0e6208e03a3cfa06a38c63b34282736fc3a3489
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3312505
Reviewed-by: Keith Short <keithshort@chromium.org>