flash_fp_mcu: Drop -kernelnext from platform name
If we're running flash_fp_mcu on a board that has a kernelnext variant,
then it's essentially the same platform, but using a new kernel. Let's
just drop the variant name so that everything else works like before and
the flashing script can be run on new kernel variant testing boards.

BRANCH=None
BUG=b:201110424
TEST=tast run $strongbad_kernelnext_dut firmware.FpFlashFpMcuHello

Change-Id: Id35f9f06b4a3ecc27891a6d22a8958cc99b36b04
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3318784
Reviewed-by: Alexandru Stan <amstan@chromium.org>