flash_fp_mcu: Find gpiochip base dynamically for strongbad
Instead of hardcoding the gpiochip base in here, let's find the gpiochip
that we care about by dereferencing the path in /sys/class/gpio/ and
checking it against a known device on the platform. This fixes this
script on strongbad-kernelnext where the base number has changed
slightly and makes this script more robust against it changing again in
the future.

We also use PLATFORM_BASE_NAME instead of PLATFORM_NAME in another place
because it doesn't matter if this is a board variant for this
conditional.

BRANCH=None
BUG=b:201110424
TEST=flash_fp_mcu on strongbad-kernelnext

Change-Id: Iaceb3cab95977cb2ebf3d699fc2e31e317403b2e
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3331672
Reviewed-by: Alexandru Stan <amstan@chromium.org>