Guybrush: Restore TCPC I/O pins after TCPC reset
Right now, we may reset the TCPC during runtime in order to get it out
of dead battery mode so that we may control the sink enable pin.
However, when the TCPC is reset then the I/O ports on it will need to be
restored in order for the Type-A ports to work correctly.

The Type-C related I/O pins should be irrelevant since the port will
be going through Error Recovery after the reset, and will be set there.

BRANCH=None
BUG=b:210802721
TEST=on dewatt, confirm A port works on a boot when the battery is
being brought out of cutoff and C0 gets reset

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I5ac57c2bed5235c94446e51975c793abcfbffb7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3345994
Reviewed-by: Rob Barnes <robbarnes@google.com>
Tested-by: Rob Barnes <robbarnes@google.com>