stm32/hwtimer32: Clear status register after counter is set
Whenever a timer interrupt occurs, the status register is checked for
overflow and the overflow status is passed to process_timers(), which
increments clksrc_high. When setting the clock to a new time, we don't
want the previous overflow state to trigger this process.

To avoid this problem, just clear status register when new value is
assigned to counter. To avoid race between updating counter and clearing
status, operation is performed when counters are disabled.

BUG=b:200828093
BRANCH=none
TEST=make -j buildall

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: Ief65ac86fcf5c164f99855cc08c96c2ceddb29ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3347657
Reviewed-by: Tom Hughes <tomhughes@chromium.org>