zephyr/dts: don't require irq for pi3usb9201
The shim driver for pi3usb9201 appears to have been designed to
gracefully handle the absence of an irq property, but they way it was
written still required its presence. The IRQ sometimes needs to be
omitted when a board shares the BC1.2 IRQ with some other device, so
mark the property as optional and fix the driver to still compile when
it's not present.

BUG=b:211693800
TEST=zmake buildall
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I293a4e88b9a2e5006141654afd9e667fd3f8ecdc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3360322
Reviewed-by: Yuval Peress <peress@google.com>