phaser: add psl code back for some sku request
Add the the PSL code to the phaser's board.c file,
and use sku id to distinguish for enter PSL or not.

BUG=b:203442963
BRANCH=None
TEST=make buildall -j
In Phaser, keep KSI1, KSI2, and KSI3 on when system in hibernate
In Laser, only KSI2(powered by H1) on when system in hibernate.

Signed-off-by: arthur.lin <arthur.lin@lcfc.corp-partner.google.com>
Change-Id: I55ee776d5fd4c500511ee14f0715c3850a4e16d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3372133
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Henry Sun <henrysun@google.com>