nissa: Remove PPC, fix task event send
Remove PPC references.
Fix task event send to charger task.
Add support for checking port count in interrupts.
Fix PPC config conditional so that other options are available.

Sometimes there still seems to be a loop when the board is restarted
with a battery and power from a servo on port 0, but this
may be a servo issue.

BUG=b:213835377
TEST=zmake configure -b nivviks; flash and run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I3fcd53f32a9fba894018538e12458ec9ae811075
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3379790
Reviewed-by: Peter Marheine <pmarheine@chromium.org>