tcpm/rt1718s: add board hook for set_snk_ctrl()
RT1718S provides 3 GPIO for universal uses, and one of the
use is making one of the GPIO a SNK enable pin.

For such case, we add a board hook for board customization, so
the board can control the pin as their needs.

BUG=b:214146359
TEST=ensure the board hook is called
BRANCH=none

Change-Id: I15955e07bfcd192ea79ab891da58021a2ec153c6
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3386981
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Eric Yilun Lin <yllin@google.com>
Tested-by: Eric Yilun Lin <yllin@google.com>