ectool: add support for the MEC LPC protocol
The MEC LPC protocol is supported in Linux' cros_ec_lpcs driver, but not
in ectool's "local" implementation. However, cros_ec_lpcs will not load
on the Framework Laptop due to an I/O port conflict[^1]. Therefore,
ectool's local implementation is required to support the MEC protocol on
kernel versions that do not contain a fix for the I/O conflict.

[^1]: https://lkml.org/lkml/2022/1/4/1160