Makefile: Switch to the Python script for CONFIG checking
This is the second attempt, fixing the problem with CONFIG_DAC and
other options. These must be left in config_allowed but ignored if
they show up as present in Kconfig.

Leave the following in config_allowed since they are defined both
as CONFIG options in ECOS and Kconfig options in Zephyr:

    AUDIO_CODEC
    DAC
    DMA
    EEPROM
    I2C_BITBANG
    PECI
    SPI
    UART_CONSOLE

BUG=b:181323955
BRANCH=none
TEST=./util/kconfig_check.py -c build/blipper/.config
   -a util/config_allowed.txt -p PLATFORM_EC_ -s zephyr/
   -I ~/cosarm/src/third_party/zephyr/main -i DAC check
See that this does not cause an error now

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I1fc7ac9cde106c61dd28b30673830407aa8c1a29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3425452
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>