zephyr: zmake: Disable traceback suppression
As discussed in the logging brainstorming meeting, traceback
suppression only makes it more difficult to fix bugs, determine the
cause of a failure, or make changes in zmake.  This is to the point
that much of our documentation has been putting -D everywhere.

While there's much work to go to get rid of -D everywhere so that we
display the right logs when we need them, we can start at getting rid
of the traceback suppression feature and making -D just set
--log-level=DEBUG.

The two options are placed in a mutually-exclusive group such that
specifying --debug/-D and --log-level at the same time is not valid.

BUG=b:217423904
BRANCH=none
TEST=unit tests pass

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I124b23c307b8b6e82edf4d5e43e2fcff82384a8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3432911
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>