COIL: Update I2C terminology
This CL replaces slave_addr_flags with addr_flags.
For context, "addr_flags" is the name is that is being used in the
caller function. E.g:
http://cs/chromeos_public/src/platform/ec/common/i2c_controller.c?l=137

So, the naming convention is the name as the rest of the functions,
and also conforms with COIL.

BRANCH=None
BUG=b:163885307
TEST=compare_build.sh -b all matches

Signed-off-by: Ricardo Quesada <ricardoq@google.com>
Change-Id: I75efe758daa8f51ae37f6ba2c67c3a7bf8ce3a8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3436449
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Commit-Queue: Ricardo Quesada <ricardoq@chromium.org>
Tested-by: Ricardo Quesada <ricardoq@chromium.org>
Auto-Submit: Ricardo Quesada <ricardoq@chromium.org>