update_release_branch: Skip merge attempt when we are in merge process
Until now, update_release_branch.py script was attempting to create
merge even when there was already merge process started, which is
obviously wrong. Due to this, it was not possible to create merge
commits with automatically generated commit message when there were
merge conflicts.

This patch introduces checking if there is merge in progress. If so,
skip merge procedure and just attempt to commit changes.

BUG=b:176500425
BRANCH=none
TEST=Use update_release_branch.py script to create merge commit.
     If there are conflicts, solve them and run the script again.
     It should create merge commit correctly.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I7f550fcae7a555ec049661e4657d84741561a5b1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3440954
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>