checkpatch: Optionally allow braces for single-statement if
This disables checkpatch complaining when braces are used with a
single-statement if.

The Zephyr coding style [1] requires braces of if, else, do, while,
for, and switch, even when not explicitly required (the body is only a
single statement).

In the interest of making code easier to upstream, we should encourage
conforming more closely to Zephyr's style for all EC code.

[1]: https://docs.zephyrproject.org/latest/contribute/#coding-style

BUG=b:218330171
BRANCH=none
TEST="checkpatch.pl --no-tree --ignore SPDX_LICENSE_TAG \
         -f zephyr/shim/src/console_buffer.c" passes

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I827528ec54a62db05b8e04504cd040bb3aa66be0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3448989
Reviewed-by: Yuval Peress <peress@google.com>
Commit-Queue: Yuval Peress <peress@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>