qcom: Swap a reliable signal to confirm PMIC power
Previously, we used the PMIC_KPD_PWR_ODL as a signal to confirm if
the PMIC is powered. In the recent QC project, the KPD_PWR_N has
internal pull-up to a PMIC's 1.1V rail. The 1.1V is under the
Vih(min). So the EC may sample it incorrectly and may lead to a
wrong decision path.

The PMIC_RESIN_L signal has internal pull-up to a 1.8V. Using this
signal is more reliable.

BRANCH=None
BUG=b:218876104
TEST=Verified the power-on and power-off sequence.

Change-Id: I9217f334f2684a8b4265a44f3b10b07412a250d2
Signed-off-by: Wai-Hong Tam <waihong@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3456706
Reviewed-by: Stephen Boyd <swboyd@chromium.org>