nissa: configure voltage divider on EC_VSENSE_PP3300_S5
The hardware has a divider on the EC monitoring pin for the PP3300_S5
power rail because neither of the NPCX9 nor IT8xxx2 ECs can measure the
full-scale 3.3V rail. Configure it for the shim code.

Zephyr doesn't support generic voltage dividers on ADC inputs, so nothing
is done for those. Zephyr code that reads the ADC will need to be
divider-aware if required.

BUG=b:220001622
TEST=nivviks: `adc` now shows EC_VSENSE_PP3300_S5 at 3293 mV rather than
     2714 mV after PP3300_S5 is turned on (`gpioset en_pp3300_s5 1`).
     nereid: `adc EC_VSENSE_PP300_S5` shows expected ~3300 mV when the
     rail is turned on.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I10f49760196e41b295674a9120da6d2eed02e9fd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3470818
Reviewed-by: Andrew McRae <amcrae@google.com>