zephyr: fan: use DT_DRV_COMPAT
Define a DT_DRV_COMPAT for named_fans and use instead of calling out the
compatible explicitly. Also replace various #if DT_NODE_EXISTS() with a
single build assert, so we stop the build if the feature is enabled but
not configured.

BRANCH=none
BUG=none
TEST=build and run on volteer
TEST=zmake testall

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I5c75ac11ee19a67faaca58857b2f93b5707d1363
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3472243
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>