drop ps2 movement packets during host commands
We were running into an issue where during boot
the host would probe the ps2 mouse with magic knock.
But the ps2 mouse would be sending movement packets
with a full buffer which would confuse the probe
due to returning movement packet data mixed with
host response data.
This caused  windows driver to not detect a valid ps2
mouse if the user was moving their finger
on the touchpad during boot.
This normally only impacts OS's where there is
no PCH i2c driver present and we fallback
to ps2 emulation. Such as windows setup.

Signed-off-by: Kieran Levin <ktl@frame.work>