zephyr: fan: rename fan structured to a data/config pair
This renames fan_status_t to fan_data and fan_control_t to fan_config,
which is one step closer to the Zephyr style for driver variables and
constants, and also avoids the enum collision and drop the awkward _t
suffix which makes it looks like this is a typedef.

This leaves struct fan_t, but that's used in other EC code so it has to
stay.

BRANCH=none
BUG=none
TEST=zmake testall

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ib0fe21778c1d9b532c2e1b08bcaaa9eeb4051deb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3497453
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>