zephyr: npcx: fix caching of FWCTRL register
Add helper functions to access the FWCTRL register. This ensures that
updates to this register are reflected in the cached value.

Without this change, a "sysjump rw" actually jumped back into the RO
image, because system_get_shrspi_image_copy() always returned that the
current image is RW.

BUG=b:218820985, b:222517990
BRANCH=none
TEST=Perform sysjump on Herobrine, observe the hash of the running image
is updated to the RW copy

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: Id9f8e8703f3141a2f8eff7aa36b1ccc7e6cbcfb8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3502648
Reviewed-by: Andrew McRae <amcrae@google.com>