zephyr: pwm: add generic-pwm-channel support to kblight and displight
On ARM platforms, the host uses the PWM host command to control the
display and keyboard backlight level. Unfortunately the current driver
does not use the stable PWM types (EC_PWM_TYPE_KB_LIGHT...), but instead
relies on using a known generic PWM channel number (0 or 1), which is
mapped differently on different boards.

To restore compatibility with the current driver, add a PWM channel
number property to the kblight and displight driver, and use it to
control the correct PWM from the PWM host command.

BRANCH=none
BUG=b:217741090, b:222516095
TEST=zmake testall
TEST=ectool pwmsetduty on volteer
TEST=run on lazor

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Reported-by: Michał Barnaś <mb@semihalf.com>
Tested-by: Michał Barnaś <mb@semihalf.com>
Change-Id: I68615eb178c7b8ea8bbbc392e4485d36b2f90ae8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3500630
Reviewed-by: Michał Barnaś <mb@semihalf.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>