nissa: enable CBI WP latch GPIO
Nereid from revision 0 and Nivviks from revision 1 have an EC output
that latches the CBI EEPROM's write protect on, compared to the
traditional system where the GSC drives that write protect signal.
Turn on the related software support for it.

The GPIO this uses is unused on Nivviks revision 0, so no special
handling is required for old boards.

Since this regenerates the Nivviks pinmap, it also moves the I2C port
definitions around due to the changes in https://crrev.com/c/3497401

BUG=b:223687410
TEST=builds
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I46c485623a5337450d03bb5e4a2021d3141fa843
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3515352
Reviewed-by: Andrew McRae <amcrae@google.com>