guybrush: Detect and recover from hard S0ix sleep hangs
When a sleep hang is detected, a wake event is sent to the host.
Sometimes this isn't enough to wake the host, aka a hard hang. When a
hard hang is detected, first try to reset the AP. If this fails,
aka a vary hard hang, force shutdown the AP.

Also detect consecutive hard hangs and skip the AP reset and go directly
to AP shutdown. This is done to avoid getting stuck in a loop and draining
the battery.

This change is only applied to RW to avoid the risk of getting a false
positive detection during recovery.

BUG=b:218892808
BRANCH=guybrush
TEST=Recover from a hard suspend hang on Nipperkin with KI
Recover from hard resume hang on Nipperkin with KI
Recover from simulated very hard suspend hang on Nipperkin
No false positives after 2500 suspend stress cycles on Nipperkin no KI

Change-Id: Idf1f010c2fcdc2b561c83845dce3cdacc24f1427
Signed-off-by: Rob Barnes <robbarnes@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3526173
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>