nissa: Workaround for ADC initial read error
It appears the ADC comparator initialisation breaks
other ADCs that have not been read, so workaround
the issue by initially reading all ADCs so that they are correctly
initialised.

Also add a check so that zero reads are detected and returned as an
error.

BUG=b:224900226
TEST=zmake build nivviks; flash and run and check ADCs work.
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Ifcbebf2ca9b9a336f10fe3659fd3452223474ed3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3528015
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>