zephyr: Add "ghost" variant of brya
Right now using brya hardware to develop ghost EC.  Initially it can
be a simple rename of brya.  Later we will change the charger chip,
which will require a couple of DTS customizations.  Eventually, this
project will need to move to not be a brya variant at some point.

BUG=b:222738712
BRANCH=none
TEST=builds (don't have hardware for testing yet)

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I3c713e12bd0685d08ed31c40bd103c2685416d78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3530958
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>