ap_pwrseq: emit a PRE_INIT event on G3->S5 transition
There is currently no event that gets emitted for the G3->S5 transition,
but legacy ARM boards use PRE_INIT in a similar way. Legacy ARM power code
emits PRE_INIT on transition from S5 to S3, but that is the transition
in which SoC power is first applied because they lack a soft-off state
similar to S5.

Since this adds new meaning to PRE_INIT where it was previously unclear,
the event is documented to be the transition from G3 to S5.

BUG=b:226215993,b:226513200
TEST=PRE_INIT is now emitted on S5 entry on Nereid
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I3ca64a309221c87583d8ae5ea8806f5ee465b0df
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3551996
Reviewed-by: Andrew McRae <amcrae@google.com>
Reviewed-by: Li1 Feng <li1.feng@intel.com>