nissa/nereid: use TCPCI mux on usb-C1
The integrated mux in the PS8745 on the type-C sub-boards is controlled
by the standard Config_Standard_Output register, but needs to be hooked
up in the EC as a USB mux. Do that.

PS8xxx also has a custom mux function for generating HPD on AUX
to the AP, which should be used: we don't use a separate GPIO for
HPD on this platform.

BUG=b:226528343
TEST=`i2c read_byte I2C_SUB_C1_TCPC 0xb 0x18` changes to 0x68 for DP
     mode and back to 0x60 when a display is unplugged.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Ic14c53a3055233b450b4d6d6262de9db9741f3ec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3555337
Reviewed-by: Andrew McRae <amcrae@google.com>