nissa: deduplicate common board_config code
The sub-board configuration had diverged between Nivviks and Nereid,
with each having some disjoint improvements. This change moves all of
the code out of board-specific board_config files and into a common
file consisting of the sum of the two boards, and the few functions that
differ between boards are moved to existing board-specific source files.

Specifically, the divergence was:
 * Nereid configured HDMI as needed
 * Nereid fixed incorrect USB-A configuration for the LTE board
 * Nivviks disabled the C1 charger task if the port was not present

BUG=b:226513609,b:214858346
TEST=HDMI board still behaves correctly on Nereid, and USB ports still
     work with the C+A board on both Nivviks and Nereid.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Iec0fac5c4fc880193a871ad302ea6752e02d8c01
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3559690
Reviewed-by: Andrew McRae <amcrae@google.com>