zephyr: Fix the override of the system workqueue
After the conversion of HOOK_INIT to SYS_INIT within the Zephyr shim
code, the EC reported a false error that the sysworkq priority was
incorrect.

Move the code that sets the system workqueue thread priority to sooner
in the initialization order.  Also move the check of the system
workqueue thread to occur after all other threads are started.

BUG=none
BRANCH=none
TEST=zmake testall
TEST=Run "kernel threads" on byra and verify system workqueue is set to
the lowest priority, but above the idle priority.

Signed-off-by: Keith Short <keithshort@chromium.org>
Change-Id: I14743cbb6643dce9e9b73584d75ec3298c8d9117
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3562002
Reviewed-by: Denis Brockus <dbrockus@chromium.org>