zephyr: brya: setup TEMP_SENSOR_POWER
Enable CONFIG_PLATFORM_EC_TEMP_SENSOR_POWER and set the pin name on Brya
(it's connected to the S5 rails enable) so that the temperature sensors
don't report max temperature when the AP is powered off.

Long term the GPIO should be referred from a named-temp-sensors
property, so add a TODO for that.

BRANCH=none
BUG=b:227761084
TEST=build and run on brya
TEST=temps on the EC console

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ie83451ee0e73dda05d7132e990e8802ca4d3e8f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3563142
Reviewed-by: Yuval Peress <peress@google.com>