jsl: Kick chipset SM if power loss in suspend
On the jasperlake platforms, our DSW_PWROK is actually generated by the
EC but it relies on the ADC interrupts to be enabled.  For power
reasons, we do not keep the ADC interrupts enabled during suspend as the
assumption is that the power rails will not be lost.  However, it seems
that that may actually occur. This would cause the chipset state machine
to be stuck in suspend when the system is actually no longer in suspend.

This commit simply has the EC do a couple of things:

- In suspend, assume that the PP3300_A power good follows the enable
- During a force shutdown, manually trigger the chipset state machine to
  re-evaluate the current signals.

BUG=b:218596390
BRANCH=firmware-dedede-13606.B
TEST=Build and flash magolor, login as guest, close lid, wait for system
to enter S0ix.  Issue `apshutdown` on EC console, verify that DUT enters
S5 and G3. Open lid, verify that DUT boots back to S0.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>

Change-Id: I2e104857851f245bf37574a8a47e415633fb9e77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3566111
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Feels: Aseda Aboagye <aaboagye@chromium.org>