zephyr: Undo changes to setup GPIO outputs.
It's likely that b/218773662 was caused by an unrelated
issue, possibly a h/w issue.
The GPIO driver does not reset the outputs except under
a full system reset, so do not set the outputs
of the GPIOs when warm-booting. The change to avoid this
for open-drain outputs is no longer required (all outputs
are now left as they were).

BUG=b:218773662
TEST=zmake testall; zmake build nivviks; flash & run.
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Iab46d79148e4f445c608e308a035c4e2938194e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3563453
Reviewed-by: Wai-Hong Tam <waihong@google.com>