ec: Add a sleep to make coverage stable
The function tc_error_recovery_run is getting called by the PD task
after the end of the test, so sometimes it appears in the coverage
report and sometimes not. Add a sleep to make sure it always has time
to run.

BRANCH=None
BUG=b:213374060
TEST=Ran usb_typec_ctvpd 50 times

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: Ibf3019e7add2522ae8e5287c6053df2a8f9ea596
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3572888
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>