zephyr: shim: decouple syswork support data from task_ctx_data
This is an alternative implementation of:

ff45017216 Zephyr: Create sysworkq entry in task shim

The original one was adding a full new task context structure for the
sysworkq task, but that was not making any use of the "struct k_thread"
field, which takes quite a bit of RAM.

This version splits out the base data on a separate structure and uses
that for the sysworkq handling. Additionally it moves the timer
structures on a separate array to avoid structure padding (struct
k_timer ends up with an int64_t).

Net saving on RAM is 232 bytes on brya and 192 bytes on nereid, the
difference is mainly due to the FPU registers.

BRANCH=none
BUG=b:223044986
TEST=build and run on brya
TEST=cq dry run

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: If2b3a53875900c26f2357a08ff3f9beb7c63c5e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3575153
Reviewed-by: Diana Z <dzigterman@chromium.org>