ec: Add sleep to emulated ztests for tasks
Sometimes the tasks don't finish starting until the test is done.
There are several tasks that are included in all tests even if they
don't use them, such as the console task, and those often don't even
start until the end of the test. This causes the number of covered lines
in gcov to change from run to run.

Add a sleep before printing Pass, as that is when the util/run_host_test
kills the test.

BRANCH=None
BUG=b:213374060
TEST=Ran accel_cal 50 times

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: I36a8b2e0812adbbb158c439aea867e1b5bc960d6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3573692
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>