test: add test for fifo pop
Test the case where the entry being removed has fewer than 2 entries
in the fifo. This would trigger us to look at other sensors to see if
timestamp spreading is still needed. The test verifies that the
correct sample is at the head of the sensor data and that the
timestamps were spread.

BRANCH=none
BUG=b:224614211
TEST=make run-motion_sense_fifo -j

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: I5771247722d0e2cc4b1765d9da2b6d66e649303d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3575698
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>