dedede: Update pp3300_a_pgood when lost
Recently, we made a change that would cause the PP3300_A_PGOOD to follow
that of the enable, but it failed to update the local variable.  This
can cause us to assert DSW_PWROK which can violate the Intel PDG
requirement which may lead to RTC corruption.

This commit simply updates the local variable in step with the logic
that was recently added.

BUG=b:218596390
BRANCH=firmware-dedede-13606.B
TEST=Build and flash magolor, boot to S0, suspend to S0ix, issue
`apshutdown` on EC console, verify that DUT shuts down to S5 and that
DSW_PWROK isn't asserted durng this process.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I420b6e0d29b17bc07b02b3598ec5c099afecfa57
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3576652
Tested-by: Aseda Aboagye <aaboagye@chromium.org>
Auto-Submit: Aseda Aboagye <aaboagye@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Diana Z <dzigterman@chromium.org>