taniks: Override NVME and EMMC FW_CONFIG bits
Taniks has an EMMC_SKU_DET strap pin which signals whether or not the
current board has EMMC or NVME storage. In the factory, the CBI images
are prepared with both NVME and EMMC masks set to enabled, but this
causes an issue in the AP firmware PCIe RP initialization because the
two devices are designed to use the same CLKREQ# pin regardless of
SKU. The FSP cannot handle this, thus we are left with trying to
override these values. Thus this patch will read the FW_CONFIG value
at init time, and if either both NVME and EMMC are selected, or
neither, then it will read the EMMC_SKU_DET pin and set the two
FW_CONFIG masks accordingly.

BUG=b:224884408
BRANCH=none
TEST=make buildall -j
TEST= verified pass on taniks MB for rework GPIOD4 emmc_sku_det.

Signed-off-by: arthur.lin <arthur.lin@lcfc.corp-partner.google.com>
Change-Id: I9c10ef52793d5c843aa65357a154b579aa5459aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3577506
Reviewed-by: Parth Malkan <parthmalkan@google.com>
Reviewed-by: Boris Mittelberg <bmbm@google.com>