zephyr: kblight-pwm: make the option dt node based
Change the Kconfig kblight-pwm option to enable automatically if a a
cros-ec,kblight-pwm node is defined.

Also move the option out of the "choice" block, as that was made for
port expander support, but the driver has since been ported to the
Zephyr PWM APIs, so once we'll have an IT8801 PWM driver available,
it'll work with the existing driver automatically.

BRANCH=none
BUG=b:217741090
TEST=compared the .config on brya

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I46ddff2613b901440162e22b89895e591435bb8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3578648
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>