corsola: re-config usb mux at DP post-config
If we plug a dongle connected with an external monitor, the
HPD level will be sending high at the first DP STATUS command.
In this case, the USB MUX then will be put into safe mode before
post-config and the previously configured mux mode in DP STATUS will be
forfeited. So, we should re-configure it again (and this is also in
the common code) if it's the selected aux-port.

BUG=b:228289594
TEST=plug a dongle with an external monitor connected, and is able to
     output.
TEST=plug C0 DP dongle, then C1 dongle, and plug an monitor to C0
     dongle, and it has output. (test aux path switching).
TEST=plug C0 DP dongle, then C1 dongle, and plug an monitor to C1
     dongle, and it has output. (test aux path switching).
BRANCH=none

Change-Id: I3ebfeee5d4a89475721b2cb2ca6050d808fc44ac
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3592414
Tested-by: Eric Yilun Lin <yllin@google.com>
Auto-Submit: Eric Yilun Lin <yllin@google.com>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>