zephyr: zmake: Add test_args to project configuration
For tests, provide a mechanism to pass extra arguments on the command
line to the test binary.  A temporary directory is provided to place
temp files.

The purpose of doing so can be seen in the modified BUILD.py for the
drivers test: the drivers test has been littering a `flash.bin` file
in the working directory, and we'd like to stop that if we can.

BUG=b:230107609
BRANCH=none
TEST=zmake test test-drivers -> no flash.bin litter

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I52e9ed68489eea46b109e0e553bd6cf26c23a64a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3602223
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Aaron Massey <aaronmassey@google.com>