nissa: don't control HDMI sub-board S5 rails
The EN_S5_RAILS input to the HDMI sub-board is not actually connected to
anything (it is enabled as long as the S5 3.3V rail from the mainboard is
provided to the sub-board), so it doesn't need to be controlled by the EC.

If the EC is asserting EN_S5_RAILS to the sub-board when attempting to
hibernate, hibernate fails because it is latched back off by the
asserted signal (anytime the AP is in S5 or higher power state). Since
hibernate should work in S5 as well as G3, we can simply stop
controlling that EC output and the sub-board will continue to work
normally.

BUG=b:226325993
TEST=Nereid sb_1 (SUB_USB_C1_INT_ODL) remains high when AP is on with
     sub-board 3 in use, and the sub-board remains powered (the USB port
     on it works).
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Idd6c0a37249d1e85a8fd0ffc763a31cee452f9cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3612982
Reviewed-by: Andrew McRae <amcrae@google.com>