power: Add new chipset_power_on API
This is a simple API for subsystems to request the chipset to power
on.  It's required so that the chargesplash state machine can request
a chipset power on upon AC connection.

Right now it's only implemented for x86 and Qualcomm as an example.
Support for other platforms can be added as required.

BUG=b:228370390
BRANCH=none
TEST=with follow-on CL, chargesplash can power on

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I73a0fa21d5aea6bc46ad23eed2b8d908d217fe0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3615483
Reviewed-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>