zephyr: Do not shutdown alt mode for sysjump
Do not shutdown USB PD alt mode before a sysjump.
When the shell console command 'sysjump' is used, the
the task ID is invalid, causing the EC the hang waiting
for the invalid task.

TCMPv2 does not require the pre-sysjump preparation any more,
since most connections now start in ErrorRecovery.
This was previously required for situations when a monitor
was connected, and the connection was required to be maintained
across sysjump restarts. Now that the connection starts in
ErrorRecovery, the monitor connection will be restarted.

Previous behaviour is maintained for backwards compatibility,
and this is only turned off for Zephyr based systems.

BUG=b:230572489
TEST=zmake build nivviks; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Ib541105a3f1a1fd8b83b6363063df10981c84fed
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3616364
Reviewed-by: Peter Marheine <pmarheine@chromium.org>