test: fix flaky test
Remove check for value that's not being deep copied. It's possible to
do this with fff but it requires a custom handler to do the deep copy.
This might be worth exploring once we enable C++ with lambdas which
will make the custom handlers much easier.

BRANCH=none
BUG=none
TEST=zmake test test-drivers

Signed-off-by: Yuval Peress <peress@google.com>
Change-Id: If454281440fdc3f5f55d7243301c512de870076f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3621871
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>