chip/npcx: Add function to poll DIRTY bit for virtual wire changes
The NCPX eSPI Module has a set of registers called VWEVSM. These
contain information about the various virtual wires connected via eSPI
to the host. One of these is a 'DIRTY' which will be set when the host
recognizes a change in the status of one of the virtual wires that the
particular VWEVSM register is responsible for.

This patch adds a function to poll the DIRTY bit for a given virtual
wire until it is no longer set, i.e. the VWIRE change has been
acknowledged by the host.

BUG=b:230366077
BRANCH=none
TEST=tested with next patch

Signed-off-by: Tim Wawrzynczak <twawrzynczak@chromium.org>
Change-Id: I87ae66c7af067bd14633edf04f6ee9dc5ffcde99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3626782
Reviewed-by: caveh jalali <caveh@chromium.org>
Reviewed-by: Boris Mittelberg <bmbm@google.com>
Reviewed-by: Jeremy Compostella <jeremy.compostella@intel.corp-partner.google.com>