zephyr: it8xxx2: Do not default GPIO groups K/L to inputs
By default groups k/l are set to input/pull down on startup
since for some variants they are not exposed on pins.
For the IT81302 they are, so do not set these as inputs
at startup by default, otherwise it breaks warm reboot
for any of these pins that are configured as outputs.

BUG=b:231727461
TEST=zmake build nereid; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I036b2ef5ec6bb38431b00644b90db44393a9384f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3631842
Reviewed-by: Peter Marheine <pmarheine@chromium.org>