fix dropped interrupt from touchpad
This fixes an issue where the ps2mouse task will sometimes
not process an interrupt from the touchpad, and the task
will get 2 interrupt events on top of each other causing
the task to get touchpad data after the touchpad has
deasserted the interrupt line.

I think the smbus controller is still getting in a state
and hitting the 100ms timeout, which happens right on top
of the touchpad 100ms interrupt retry logic. So the
touchpad task gets a 2nd interrupt while it is servicing
the first interrupt. But the touchpad only responds
one time

Signed-off-by: Kieran Levin <ktl@frame.work>