tests: Set TCPCI revision in suite usb_attach_5v_3a_pd_source
The test suite usb_attach_5v_3a_pd_source contains two flaky tests
(test_power_info and test_typec_status) that fail when the suite is run
in isolation or out-of-order.

Abe noticed that the test doesn't set the TCPCI revision, which we need
to be higher than the default of 1.0. These tests were previously
relying on other test suites to make this update.

This change sets the TCPCI version in the board config and emulator
(both are required and must agree) in the suite's setup function. It
may be required in other suites as well, but those will get handled on
a case-by-case basis as test shuffling reveals more flaky tests.

BRANCH=None
BUG=b:232431388
TEST=zmake test test-drivers

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I55f1ad46c3dfb370f43a6f80cd93913a95335f9a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3646049
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>