brask: Use GPIO and host event when generating MKBP events
The power button of the brask is passing via the MKBP event.
After landing the http://crrev/c/3570523, the deptcharge would
wait for the EC_PCH_INT_ODL interrupt to trigger the MKBP event.
As a result, we need to follow this http://crrev/c/3544873 to
add the MKBP interrupt to avoid the missing power button key
event in the recovery mode.

BUG=b:209707579
BRANCH=none
TEST=test_that -b brask ${IP} firmware_ECPowerButton

Change-Id: I59879d47399768ea1e4165ada71eafe6f64f1842
Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3647088
Tested-by: ChiJen Chen <chi-jen.chen@quanta.corp-partner.google.com>
Reviewed-by: Subrata Banik <subratabanik@chromium.org>