nissa: Add braces for variable declarations.
With riscv64-zephyr-elf-gcc variable declarations can't come after the
switch label, and this is a C99 restriction, so it is strange that it
works in any compiler.

Add braces around the code in the switch statement, so that the variable
declaration follows a { instead of a label.

BRANCH=None
BUG=None
TEST=git cherrypick to gitlab and built there

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: Ic2adbe92c76fc299f06dc379a930b98e50d52aaf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3652376
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>