zephyr: drop ioex_init for Zephyr ioex drivers
The ioex_init function has to reconfigure the ioexpander chip. For the
Zephyr ioexpander driver, which is treated as a normal GPIO driver,
means we have to configure every single pin from the ioexpander chip.

Use the gpio_reset_port function for that purpose instead of ioex_init
which bases on the ioex_config array(takes an index of the array as
argument).

This change allows reconfiguring ioexpnader pins e.g. after the
ioexpander chip reset which is done during TCPC init for Brya board.

BUG=b:231086773, b:208181774
TEST=zmake testall && Make sure PD works on ports 0 and 2 on Brya
BRANCH=main

Signed-off-by: Dawid Niedzwiecki <dn@semihalf.com>
Change-Id: I8c446a4b35b03b9ed9574c6572a0daa80a9a7dee
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3655999
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>