zephyr: shim: add single charge task support
Add an option to run the charger code in a single task.

This is enabled by default in Zephyr builds, but explicitly disabled for
a couple of platforms that currently have task dependent board code.

Should be enough to start testing the configuration, will work on
handling the remaining boards (corsola and nissa) later.

BRANCH=none
BUG=b:226411332
TEST=make buildall
TEST=zmake testall
TEST=cq dry run
TEST=run on Brya, connect/disconnect usb devices

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I75a63f7b0a9545e6c824114de7f81b71924e0789
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3663748
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>