zephr: Handle shutting the shell down via an asynch signal
To communicate with the GSC, the EC needs to shut down the shell
and take control of the UART.

Before shutting down the shell, the uart interrupts are disabled.
To shut the shell down, a KILL signal is sent to the shell,
and a callback is used to coordinate when it is shut down.
Console output must then go through printk rather than
shell_fprintf, otherwise there may be a deadlock during the
shutdown (since zephyr_fprintf holds the shell context lock,
and may be waiting for the output to complete). This stops
the shell from processing the KILL signal.

Add a flag that indicates the shell is to be shut down so
that the deadlock can be avoided.

Also remove calls to shell_process, which is called
from the shell mainline under the context mutex.
If input is to be processed, it should be done within the
context of the shell triggered by the SHELL_SIGNAL_RXRDY signal.

BUG=b:233309416
TEST=zmake build nivviks;flash & test
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I2187ac470c6e10b54540f357690f79aaa0576271
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3670625
Reviewed-by: Keith Short <keithshort@chromium.org>