common: Fix host command print formatting
Modify the format of how the host commands are being printed. Change the
zero paddings from 2 to 4 to conform with the command, which is defined
as a 16-bit value. The previous method will not result in incorrect
output for commands requiring more than two digits.

BUG=b:231975323
BRANCH=None
TEST=./test/run_device_test.py --board bloonchipper
TEST=make buildall

Signed-off-by: Firas Sammoura <fsammoura@google.com>
Change-Id: I4dd5507096a5a02f9554a164cae610683aebfec0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3705694
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Tom Hughes <tomhughes@chromium.org>