chip/stm32: Allow controlling SPI chip select via USB
uServo and other similar EC devices allow a USB host to instruct them
to perform SPI transactions.  Until now, the EC devices have asserted
SPI chip select only for the duration of a single SPI transaction.

HyperDebug is a EC-based generic debugging aid, which aims to be able
to communicate with many SPI devices, some of which use protocols that
require chip select to remain asserted across several SPI
transactions.

This change adds a new USB command to allow the USB host to instruct the
EC device to assert or deassert chip select.  If chip select has not
been explicitly asserted before performing a SPI trnasaction, it will be
asserted for the duration of that transaction, and then released, as was
the behavior before.  If chip select was explicitly asserted, it will
remain asserted during and after any SPI transaction until explicitly
deasserted.

BUG=b:192262089
TEST=Used HyperDebug to interact with AndreiBoard via SPI
BRANCH=none

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Icb24f67ae1f3683c0c8d054dc3ea34fa32c2ada5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3698582
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>