ps8743: Fix mode to apply power optimization and value to write
Apply the power optimization if USB mode is on and DP mode is off. The
check of these modes are achieved by checking the Misc register.
Correct value to write to the mode register. As the optimization only
applies when only USB2 devices are connected, and should not turn on USB
mode if device is removed during S3, create enum and record status with
internal variable, then restructure the logic for suspend and resume.

BUG=b:229810117
TEST=Check the operation mode register (0x00) entering/leaving S3:
     (1) USB 3 device is connected, register value not changed with
         different type c side.
     (2) USB 2 device is connected, only the USB mode bit is unset at
         S3, and set in S0.
     (3) USB 2 device is connected in S0, and remove the device when in
         S3. USB mode bit is unset after back to S0.
     (4) Connect a A->C adapter in S0, then insert USB 2/3 device in S3.
         The USB mode bit is set after leaving S3. `lsusb` can identify
	 the device under USB 2/3 bus.
     (5) USB mode bit is always unset if no device is connected.
     (6) type c display is connected, register value not changed.
BRANCH=None

Change-Id: Ic85ee805ffea26bfa3fc5004c3d319cefcf2e68e
Signed-off-by: lschyi <lschyi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3702073
Commit-Queue: Sung-Chi Li <lschyi@chromium.org>
Reviewed-by: Eric Yilun Lin <yllin@google.com>
Tested-by: Sung-Chi Li <lschyi@chromium.org>