zephyr: EEPROM_CBI_WP does not imply BYPASS_WP_CHECK
Some platforms allow the EC to control CBI WP by latching the hardware
WP signal on with external hardware when the EC asserts a selected GPIO,
enabled with the EEPROM_CBI_WP option. Because this should be possible
to use when the system is not force-unlocked by config, remove the
kconfig select.

This is consistent with the legacy configuration where some boards enable
EEPROM_CBI_WP but not BYPASS_WP_CHECK, and not doing this causes build to
fail because config.h only allows BYPASS_WP_CHECK to be enabled when the
system is force-unlocked.

BUG=b:235440351,b:235425819
TEST=PLATFORM_EC_BYPASS_CBI_EEPROM_WP_CHECK is now disabled on Nereid
     if the system is not force-unlocked.
BRANCH=none

Change-Id: I5c67fcb9cd564622f4f4593f358b993c4bc159b6
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3702878
Reviewed-by: Andrew McRae <amcrae@google.com>
Reviewed-by: Denis Brockus <dbrockus@chromium.org>