util: Move stdlib declarations to builtin directory
The "builtin" directory is EC's copy of the C standard library headers.
Move the declarations for functions that are provided by the standard
library to the "builtin" directory. This change makes it easier for
future changes to optionally build with the C standard library instead
of the standalone EC subset.

BRANCH=none
BUG=b:172020503, b:234181908
TEST=./util/compare_build.sh -b all -j 120
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I4822e49677e0c0e1b092711b533f2aaa762cb4e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3712033
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>