tree: Move printf declarations to stdio.h
The "builtin" directory is EC's copy of the C standard library headers.
Move the declarations for printf functions that are provided by the
standard library to the "builtin" directory. This change makes it easier
for future changes to optionally build with the C standard library
instead of the standalone EC subset.

BRANCH=none
BUG=b:172020503, b:234181908, b:237823627
TEST=make buildall

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I4a5b4f8b98b972e86c4cca65d4910b5aa07ec524
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3712034
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>