sm5803: ensure clocks are not in LPM when enabling
If the charger's internal clocks are running at their reduced
frequency when switching is enabled, it is possible for its internal
control loops to respond slower than is intended and cause incorrect
behavior or possibly even damage to the charger. This change ensures
that the clocks are running at full speed before enabling either
sink or source mode on the charger, providing strong runtime assurance
that the charger should not be damaged.

BUG=b:230712704
TEST=Bus-powered devices continue to work normally when connected to
     Nereid, and the battery can be charged.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I656605921648b72c1da7394012dc7a9a32397895
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3711514
Reviewed-by: Andrew McRae <amcrae@google.com>