TCPMV2: Notify AP of cable identity when connected to non-PD partners
If a non-PD sink is connected to a USB-C port using a PD cable, the EC
will discover the cable's identity. But, once it fails partner discovery
it will set state to PE_SRC_DISABLED without notifying the AP about the
cable. This CL updates the PE to notify the AP of cable discovery if the
SOP' identity was discovered before source discovery failure caused by a
partner not supporting USB PD.

BUG=b:235534422
TEST=make try_build_boards and make runhosttests.
BRANCH=None

Signed-off-by: Jameson Thies <jthies@google.com>
Change-Id: I1e7d042cdf0fb5194c3334c8a4f8680579fe2e68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3717919
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>