flash_ec: improve diagnostics output and stop earlier on gsc servo errors
In case there is a problem for servod connecting over CCD the script
keeps going using 'neither' as the CCD device type, which makes no
sense.

Quit the script as soon as the device type is reported as 'neither',
also show the device type in the log when connection is successful.

BUG=none
TEST=verified that the script exits early and properly reports the
     device type problem when servod fails to properly connect over
     CCd.

Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Change-Id: I52d227d98a434ff9ddbd0c7aad3111a391feb1fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3718839
Reviewed-by: Jett Rink <jettrink@chromium.org>