espi: Use VW PLTRST# instead of ESPI bus reset for reset hook
The legacy EC-OS uses the virtual wire signal PLTRST# to
drive the HOOK_CHIPSET_RESET hook. Align the Zephyr espi shim
to also use this VW signal instead of using the espi bus reset
callback.

This causes the reset hook to be called in warm boot as
well as cold boot.

BUG=b:236902047
TEST=zmake build nivviks; flash & reboot
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Ib9bb2cf60d4efa895158527163d4d86cdd7e5f8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3715184
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>