WORKAROUND: delay 3 seconds to disable BB retimer power
When the system resumes from S0ix, it will do the HDMI handshake.
But if EC disable the BB retimer power at this moment, it will cause
the handshake is failed.

EC can not know when the system actually resume from S0ix, so add the
workaround to delay 3 seconds to disable BB retimer power, and cancel
the delay function when EC want to enable the retimer power.

Signed-off-by: Josh Tsai <josh_tsai@compal.corp-partner.google.com>