tests: Fix chargesplash within-suite-OD failure
The chargesplash test helper function force_chipset_off restarts the
chipset state in order to test the lid opening triggering the
chargesplash screen. However, not enough ramp-up time seems to be
happening afterwards for GPIO interrupts to properly fire.

Abstract the helper as a common utility function with an additional
virtual sleep inserted at the end so that future test-writers do not run
into the same problem.

BRANCH=none
BUG=b:236726670
TEST=zmake test-drivers
TEST=zmake test-drivers-chargesplash shuffled with shuffle_test.sh

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: Idbcfd65a7ec66ebbd759d1a85447127d570d0099
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3739317
Reviewed-by: Yuval Peress <peress@google.com>