it8xxx2: disable eSPI inputs when in G3
eSPI CS# goes low when the AP is off, which prevents the
EC from entering deep doze to save power. Disable the eSPI
pad on G3 entry and re-enable before exiting G3 to enable
EC power saving.

This optimization is expected to be relevant to all boards that use eSPI
to manage AP power with IT8xxx2, so the code is located such that it will
be included for every board that has appropriate configuration.

BUG=b:237717730
TEST=Nereid still boots from G3 and returns to G3 correctly; debug logs
     indicate eSPI pad control is running.
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Ia2f245b0fc84cdddf28471a13a69e73b6188f932
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3742840
Reviewed-by: Andrew McRae <amcrae@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>