test: Remove TEST_CHECK macro
The TEST_CHECK macro returns a value (from the calling function) on
either success or failure. This differs from all other TEST_* macros in
test_util.h, which only return on failure. Returning on failure is
appropriate and allows short circuiting a test on the first failure but
returning on success results in short circuiting after the first
successful check and bypassing subsequent checks. This behavior
is somewhat confusing and easy to miss during review.

BRANCH=none
BUG=b:238120333
TEST=make runhosttests
TEST=./util/compare_build.sh -b all -j 120
      => MATCH
Signed-off-by: Bobby Casey <bobbycasey@google.com>
Change-Id: I3777b427aa5e20a91689f86fc37daadffacf27f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3748830
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>