zephyr: switchcap power off delay
Some switchcaps do not have power good signal which can give an
indication to EC when system is power on or off. Due to lack of this
feedback path, there isn't any way for EC to determine when the
switchcap can be turned on reliably after power off. This patch
introduces a delay parameter which can be used to wait for a specified
amount of time, determined experimentally before turning on switchcap
after switchcap is turned off.

BRANCH=None
BUG=b:212222920
TEST=Tested image on hoglin board for following after adding delay
1) dut-control warm_reset:on; dut-control warm_reset:off
2) quick power off and power on from ec console
In both the cases device can boot up without complaining AP POWER NOT
READY.

Signed-off-by: Kshitiz Godara <kgodara@qualcomm.corp-partner.google.com>
Change-Id: I97be7fbde7261a408c530710972770f7c66609d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3750958
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>