ec: Enforce black in presubmit
Enforce black, isort formatting in presubmit for all python
files.

Do not enforce flake8, because it has errors on most of the files, and
it seems to be basically the same as pylint, which is already run on
pre-submit.

Do not enforce it in the zmake unit test script.

Do enforce the formatting in the firmware_builder.py script, and do it
early for faster failures in CQ runs.

BRANCH=None
BUG=b:238434058
TEST=None

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: If3c42b2af41fd2e68accbe2867999dc931e88872
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3749243
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>