printf: Add snprintf_hex_buffer
This function replicate the behavior of the non-standard hex buffer
format "%ph" used in EC and make it possible to replace "%ph" by
printing the bytes into a buffer and then printing the resulting string
with "%s".

A followup commit will replace usage of "%ph" with these functions.
Using standard format specifiers makes it easier to switch between the
"builtin" EC standard library and the C standard library provided by the
toolchain (or Zephyr).

BRANCH=none
BUG=b:238433667, b:234181908
TEST=make run-printf

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I46135b25c84abe4ca7fb16ebb78af535622f50c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3756179
Reviewed-by: Scott Collyer <scollyer@chromium.org>