test: Fix race-condition in test_set_chipset_to_s0
The test_set_chipset_to_s0 sets the smart battery's capacity to 75%
before attempting to set the chipset in S0. However, the charge state
task may not reread battery capacity information in time before the
attempt to power on. This results in the power state remaining in G3
causing a safety check failure in the test utils code. This was revealed
by shuffling the test order execution.

Make test_set_chipset_to_s0 wait for the charge task to reread the
parameters before initiating a startup to S0.

BRANCH=none
BUG=b:233103160
TEST=zmake test test-drivers
TEST=utils/shuffle.sh

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: Ia2e948498710061f6757235470da59a61acbf66c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3761217
Reviewed-by: Yuval Peress <peress@google.com>