run_device_tests.py: Fix Pylint error consider-using-with
Pylint recommends using with for the allocation of the
ThreadPoolExecutor object.
The script uses the non-default shutdown() argument wait=False to try to
avoid to hang at the end of the test execution. However, the additional
argument does not actually prevent the hang from happening and can be
removed.

Fix the Pylint error using "with" as recommended.

BUG=b:239100048
TEST=cros lint test/run_device_tests.py
BRANCH=none

Signed-off-by: Andrea Grandi <agrandi@google.com>
Change-Id: Ia645221fa3630dfcc4fa9d9d1d1581fdb35797ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3766500
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-by: Bobby Casey <bobbycasey@google.com>