pwrseq: gpio: Use output state for GPIO get on outputs
When retrieving the state of an output GPIO, do not read
the physical level on the pin, but instead read the configured
state (by reading the GPIO configuration).

Open drain outputs sometimes use low voltages, and so
reading the physical level on the pin may indicate a
0 for an output that has been set to 1.

The stored configuration of the GPIO does not include
the GPIO_ACTIVE_LOW flag, so this is explicitly checked, and
the state inverted if necessary.

BUG=none
TEST=zmake build nivviks; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Idb2cb1c6b1b0902189e1fd1471cf9ca30c2711dc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3769706
Reviewed-by: Peter Marheine <pmarheine@chromium.org>