sm5803: prevent use when PD max voltage exceeds 15V
Silicon Mitus have recommended that VBUS never exceed 15V on boards using
the SM5803 because spikes from a nominal 20V VBUS may damage the charger,
even when there is a protection device in front of it. To prevent
accidental misconfiguration, make it a compile-time error to enable the
SM5803 driver if VBUS greater than 15V is permitted in device policy.

BUG=b:230712704
TEST=build fails if CONFIG_SM5803=y and PD_MAX_VOLTAGE_MV > 15000
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: Ie4438f1793700b4dfd2eb6f53440c8ad785f4c3a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3778901
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>