zephyr: gpio: Return configured value of output rather than pin value
When getting the value of an output GPIO, use the configured output
of the pin rather than the pin value itself.
The configured output is the value that was last set on the GPIO output,
whereas the value of the pin may not reflect this setting (i.e the
pin may be open-drain with a low voltage, or the GPIO may be an
output only GPIO).

BUG=b:236094811
TEST=zmake testall; zmake build nivviks; flash & run
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: Ic4a551ffa3c1d2e5e02e22a26645071de3c2395e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3782022
Reviewed-by: Keith Short <keithshort@chromium.org>