TCPMV2: Handle TBT mode entry
In the case of TitanRidge based TBT devices, for TBT mode entry
under AP driven mode, device must initially enumerate as a DP, when
user is not logged in. On user logging in, DP mode must
be exited and device must enter TBT mode upon AP request. For
the successful TBT mode entry, it is required that the state machine
of DP alternate mode must be ready.

BUG=b:240128813
BRANCH=none
TEST=Tested TBT mode entry on skolas with TR based TBT devices

Signed-off-by: poornima tom <poornima.tom@intel.com>
Change-Id: Idbfb9336698392e1c69b4636267a8962b10fc9cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3784490
Reviewed-by: Vijay P Hiremath <vijay.p.hiremath@intel.com>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>