rex: Use VW for sleep S3 and enable EC_SYSTEM_UNLOCKED
This CL fixes an issue where the config option was set to use VW for
sleep S3, but the power_signal node did not reflect this.

In addition, CONFIG_PLATFORM_EC_SYSTEM_UNLOCKED was added and
CONFIG_PLATFORM_EC_BRINGUP was removed to allow for the AP to boot up
automatically.

BRANCH=none
BUG=b:240434243
TEST=zmake build rex

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I83c939d5637736a7460f8f5388ebcaddf389bd35
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3823913
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Madhu P <mparuchuri@google.com>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Li1 Feng <li1.feng@intel.com>
Commit-Queue: Scott Collyer <scollyer@chromium.org>