8042: Rename I8042_CMD to ATKBD_CMD and move to a different file
These are actually AT Keyboard commands, and not specific to the i8042
controller. Using the correct prefix will make it easier to understand
which device we are interacting with. I also moved the definitions into
a different file to make it clear that there is a separation between
the two protocols.

BUG=b:242886255
BRANCH=none
TEST=build guybrush

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: I160f1d71edbce80606d8bcb01dbeb0d2192cfa93
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3835508
Code-Coverage: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Yuval Peress <peress@google.com>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>