zephyr: twister: Allow overriding toolchain options
Allow overriding the TOOLCHAIN_ROOT and ZEPHYR_TOOLCHAIN_VARIANT
environment variables, as well as the --gcov-tool CLI parameter. This is
necessary for running tests with GCC outside of the chroot, such as in
Gitlab.

By default, the usual LLVM toolchain is used and there is no impact on
users.

BUG=b:242067249
BRANCH=None
TEST=./twister -v -v

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I2369c251ac43d05bc0246d8e1b35e1bc4a28b783
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3831536
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>