8042: Abort test runner if the test gets into a bad state
If we can't propagate an error code, we can't fail a test. The next
best thing we can do is abort the test runner. Since tests have the
ability to leak state, it's possible for an earlier failing test to
break a later test. By aborting we prevent getting flooded with false
failures.

BUG=b:242886255
BRANCH=none
TEST=run unit tests

Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Change-Id: Iebdeadeeaa584a65da5329858aa24abfa6e76b03
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3846284
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>