util/comm-dev.c: Fix passing wrong parameter to ioctl
The method ec_command_dev_v2 re-calls ioctl again if the first attempt
fails and the errno is EAGAIN. There is a typo that the pointer of
pointer of data block is passed to ioctl. This may lead to segmentation
fault or, if the driver writes malicious data, arbitrary code execution.

BUG=b:243507497
BRANCH=None
TEST=None

Change-Id: Iece1ff89f9253d976e396d46c0e98f0332016bf6
Signed-off-by: Wen-Ting Tseng <wdzeng@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3850562
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>