rex: Fix tcpc_get_alert_status function
There was an error in tcpc_get_alert_status where the pin level wasn't
being read. This CL adds gpio_pin_get_dt() so the gpio pin level is
going to be read and removes the macro that was being used and was
only testing if the node existed.

BRANCH=none
BUG=b:243482165
TEST=zmake build rex

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Ife8aa4ed08c2be43776e12eeb9e28a9a04036bfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3855164
Reviewed-by: Vijay P Hiremath <vijay.p.hiremath@intel.com>
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Madhu P <mparuchuri@google.com>