core/cortex-m: Disable C++ exceptions
__exidx_start and __exidx_end indicate the start and end of the
exception index table, which is described in more detail in "Exception
Handling ABI for ARM Architecture":
https://github.com/ARM-software/abi-aa/blob/60a8eb8c55e999d74dac5e368fc9d7e36e38dda4/ehabi32/ehabi32.rst#54the-object-producer-interface

__exidx_start and __exidx_end are referenced in libunwind, so they need
to be defined. If __exidx_start and __exidx_end are equal, libunwind
will interpret the exception index table as empty, which means the
exception can not be caught, and std::terminate() is called, which is
what we want. The test in https://crrev.com/c/3932262 verifies this
behavior.

See
https://github.com/llvm/llvm-project/blob/b3c5da73b71908c59b12ef0feadfde56bad222b5/libunwind/src/AddressSpace.hpp#L515-L522

BRANCH=none
BUG=b:234181908
TEST=make buildall

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I8551ca94fce16ec34770f9204282af14667e7c43
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3857943
Reviewed-by: Andrea Grandi <agrandi@google.com>