Zephyr test: Add test for USB_PD_CONTROL
Add a test for EC_CMD_USB_PD_CONTROL.  While it's planned to be
deprecated in favor of the TYPEC_* commands, it will continue to be used
by the kernel until the time when all released boards are supporting the
TYPEC_* commands.

Test only the information retrieval the kernel uses, as the other
features produce non-deterministic command returns and are discouraged.

BRANCH=None
BUG=None
TEST=./twister -T ./zephyr/test

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I3e85fe8bbd3554e4c363c0abd717a08d78cfd729
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3858393
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>