test: Add test for get ikm failure when seed not set
Add a test for get_ikm when the fp tpm seed is not set. Change the
declaration of get_ikm function inside fpsensor_crypto.c to
test_export_static such that it can be visible in the test library.
Declare get_ikm as an extern function inside the test/fpsensor_crypto.c
file.

BRANCH=None
BUG=b:242720240
TEST=make run-fpsensor_crypto
TEST=make runhosttests

Signed-off-by: Firas Sammoura <fsammoura@google.com>
Change-Id: Ibd4993bbeb6e8d166b3d0307fd1b3c90c55cc264
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3862148
Reviewed-by: Bobby Casey <bobbycasey@google.com>