TCPMv1: Remove TBT and USB4 support
TCPMv1 has not been used for TBT or USB4 for some time. Remove this
support to make it easier to refactor TBT and USB4 code.

LOW_COVERAGE_REASON=TCPMv1 is known to be buggy but is deprecated; its
behavior is hard to describe but should not be changed.
BUG=b:244218024
TEST=make BOARD=careena; DP monitor behavior same as ToT
BRANCH=none

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: Ia02ce7a2500b83e3147ec5737134a009a59ea288
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3887961
Reviewed-by: Diana Z <dzigterman@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>