ps8xxx: Add caching to ps8xxx_get_chip_info
During boot, depthcharge uses EC_CMD_PD_CHIP_INFO to check if the PS8XXX
firmware needs to be updated. On nereid, this commands takes ~45 ms,
which affects the AP boot time.

Add caching to avoid reading registers on every call, specifically the
fw_version register and the chip-specific registers used to disginguish
between similar parts (e.g. reg_id for PS8815 vs PS8745). This is
similar to the caching done by tcpci_get_chip_info().

On nereid, the EC_CMD_PD_CHIP_INFO time is reduced from ~45 ms to ~4 ms.

Note, the reason the command is so slow is that reading the reg_id and
fw_version registers causes a chip reset, so the chip is re-initialised
in the middle of the host command (b:244386509). It's not clear whether
this is expected. If this can be fixed, then we might choose to revert
this, depending on how long the register reads take without the chip
reset.

BUG=b:242935354
BRANCH=None
TEST=Check that `ectool pdchipinfo 1` still returns the correct info.
Also add logging to check that cached info is used when live = 0, and
not used when live = 1.

Change-Id: I71dc0849270dbc729336e30d6264f7fea6898618
Signed-off-by: Reka Norman <rekanorman@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3864596
Reviewed-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>