test: Add test for read_match_secret resp error
Add a test for fp_command_read_match_secret when the number of matched
finger is within the allowed range of [0 5[, the state deadline value is
valid, the state is readable with the correct matched finger number, and
and a trivial positive_match_salt. The test should fail with the error
message EC_RES_ERROR.
Add a trivial mock value for the positive_match_salt and a default fake
value for the positive_match_salt in the fpsensor_state_mock library.

BRANCH=None
BUG=b:242720387
TEST=make run-fpsensor_state
TEST=make runhosttests

Signed-off-by: Firas Sammoura <fsammoura@google.com>
Change-Id: I9af4cb13a1b6f63049205cad9684050d5eb1f885
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3872553
Reviewed-by: Bobby Casey <bobbycasey@google.com>