zephyr: Remove boolean operator
Use the address of the host command instead of trying to force the
value to be unused. The linker will remove the unused __remove_
variable, and also the static function. At least it appears to.

GCC 12 didn't like using 0 && func as an expression because it evaluated
to a const.

BRANCH=None
BUG=b:245562553
TEST=./twister --coverage && zmake build -a
TEST=find twister-out/ -name zephyr.exe -print | xargs nm -A | grep __remove

Signed-off-by: Jeremy Bettis <jbettis@google.com>
Change-Id: Ifaa7b1593a808070da54d2c830a47d1b8893a87f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3885224
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>
Tested-by: Jeremy Bettis <jbettis@chromium.org>