test_util: Add assertion inside weakly-linked run_test()
The weakly-linked definition of run_test() should never be executed.
When it does, it's usually a symptom of an issue with the build of the
real test.

Add an assertion to make the error condition easier to detect and
resolve.

BUG=b:243700149
TEST=Run host test with bad definition of run_test()
     =>Error message is printed on console.
BRANCH=none
LOW_COVERAGE_REASON=New function is only called from invalid builds.

Signed-off-by: Andrea Grandi <agrandi@google.com>
Change-Id: Iaf6c8377618944efc5ad4d4557ed4c231c09be09
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3887968
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Bobby Casey <bobbycasey@google.com>
Reviewed-by: Tom Hughes <tomhughes@chromium.org>