usb_serial: fix console.py handling of --serialno
This was trying to pass the optional "langid" arg to usb.util.get_string(),
but was doing so as a positional argument, in the wrong position.

This bug was papered over by a blanket "except:" clause that would fall
back on not specifying langid.  With older PyUSB versions an exception from
the incorrect args would be triggered, thus the fallback was always used.
PyUSB now returns garbage for the serial number from the incorrect args, so
the serial number comparisons stopped working.

The langid being specified was 256 / 0x100, which is in a reserved block,
not specific to any real language. The get_string() invocation returns
the correct serial number with or without that langid specified (using
the args correctly), and servod never specifies langid in these lookups,
so this removes use of langid entirely to match servod behavior and remove
need for the try/except fallback.

This also fixes another blanket "except:" to specify the exception it's
actually intending to catch (IndexError).

Finally, this adds printing of default option values in the command line
help output, so that it's clear --device has a default value and will
usually need to be specified too when using --serialno.  (We should
change or remove the default and make --device optional when using
--serialno, but I'll leave that for another patch.)

BRANCH=none
BUG=b:243562728
TEST=With multiple servos plugged in, including multiple servo_v4p1,
tested opening consoles via --serialno by itself, --device by itself,
and the two options in conjunction.

Signed-off-by: Matthew Blecker <matthewb@chromium.org>
Change-Id: Idd731ce76c9ed9a2e067bca11ede42b782a94669
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3902449
Reviewed-by: Ziting Shen <zitingshen@google.com>