Makefile.toolchain: Use correct pkg-config with libftdi
The EC makefile distinguishes between building for the build machine
(BUILD) and building for the platform that the EC is running on top of
(e.g., the Chromebook's AP). PKG_CONFIG was missing this distinction,
which causes the ebuild to fail when cross-compiling for ARM chromebooks
(BUILD=x86_64, HOST=arm64).

BRANCH=none
BUG=b:246424843
TEST=CQ passes

Cq-Depend: chromium:3916417
Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: Idc107267b52f55129269851b114b09649b9b3778
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3914139
Reviewed-by: Edward Hill <ecgh@chromium.org>