rex: Change GSC packet mode signal to push-pull
ec_gsc_packet_mode was being configured as an open-drain signal, but
this needs to be a push-pull on the EC to work properly with the GSC.
This CL makes that change. Note, there is an issue opened to resolve
the Rex EC GPIO spreadsheet so that manual edits won't be required for
generated.dts. This CL is being submitted to unblock validation while
that issue is being resolved (b/244441996).

BUG=b:248775521
BRANCH=none
TEST=zmake build rex
Verified that EC is now communicating with the GSC
22-09-27 09:49:56.013 [0.416400 VB Verifying hash]
22-09-27 09:49:56.043 [0.453200 VB Received 0xec07]
22-09-27 09:49:56.046 [0.454000 VB Hash mismatch]
22-09-27 09:49:56.050 [0.454700 VB Enable USB-PD]
22-09-27 09:49:56.053 [0.455300 VB Exit]

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I98fd59f3ac728981f6541aeaac997341a381ccbd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3922461
Tested-by: Scott Collyer <scollyer@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Commit-Queue: Scott Collyer <scollyer@chromium.org>