include/common: Use noreturn attribute in test and non-test build
In the test build, the "test_mockable_noreturn" helper adds the
"__attribute__((weak))" attribute to allow overriding functions. It
should still keep the "noreturn" attribute to be consistent with the
non-test build. Otherwise, a function declared with "noreturn" that
calls one of the functions that uses "test_mockable_noreturn" will fail
to compile with:

error: function declared 'noreturn' should not return
[-Werror,-Winvalid-noreturn]

BRANCH=none
BUG=b:234181908
TEST=./util/compare_build.sh -b all
     => MATCH

Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Change-Id: I6ebfc78b7b07e19f9305a22fbfeccf62b4000667
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3933258
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>