skyrim: Remove Cr50 reset workaround config
This commit removes `CONFIG_PLATFORM_EC_BOARD_RESET_AFTER_POWER_ON` from
Skyrim. This config enables a brief delay as a workaround for a Cr50
issue that triggers a second reset of the EC shortly after it is reset.
Ti50 does not have this issue and this delay prevents GPIOs from being
configured before the jump to RW. RW does not attempt to configure
GPIOs again as it assumes that they were correctly configured by RO. The
misconfigured GPIOs cause the TCPC to be kept in reset which ultimately
causes a boot loop.

BRANCH=none
BUG=b:247543478
TEST=Ran on skyrim, verified that EC no longer boot loops.

Change-Id: I45562eba0a6bca2866ac173355765dd97e87d383
Signed-off-by: Robert Zieba <robertzieba@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3933588
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Diana Z <dzigterman@chromium.org>