test: Make zassume in util zassert
The zassume macro loudly fails a build when it fails in a
setup/before/teardown. However, it does not fail the build loudly when
it fails in a test.

Temporarily make all zassumes in util zasserts until zassume failures in
a test fail the entire test.

BRANCH=none
BUG=b:251281997
TEST=/twister --clobber -i -s zephyr/test/drivers/drivers.default

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I8545ce8db5e94e997ec70f25283d7de9c7b75237
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3937223
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>