ec: Don't declare forward decls as weak
If a function in a header is marked as weak (test_mockable), that tells
the linker that it is ok for the symbol to be missing. We never want
that.

If a function implementation is marked as weak, that tells the linker it
is ok to use a non-weak symbol instead of this one. This is what we want
when we are trying to override functions.

Remove all test_mockables from headers, except for one inline function
that I wasn't sure what to do with.

BRANCH=None
BUG=None
TEST=make -j$(nproc) runhosttests buildall && zmake build -a && \
  ./twister -v -i --clobber
LOW_COVERAGE_REASON=Added stub get_ap_reset_stats() for tests that
don't care about that function, and it's always unused.

Change-Id: Ic00a897131c93ef134544f26a8e368ce1f78a5de
Signed-off-by: Jeremy Bettis <jbettis@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3935720
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Tristan Honscheid <honscheid@google.com>
Tested-by: Jeremy Bettis <jbettis@chromium.org>
Reviewed-by: Tristan Honscheid <honscheid@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>