test: host_sleep sleep_start_suspend infinite time
Add a test that weakly verifies a call to sleep_start_suspend() with an
infinite timeout adds a hook with infinite deferred time meaning it
never gets executed and call chip-specific handlers. This is a weak
verification because it cannot prove the deferred call would never be
executed.

BRANCH=none
BUG=b:252887178
TEST=twister -s zephyr/test/drivers/drivers.power_host_sleep

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: Iee1b59096248c83fe6b513ec327f3309d1954772
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3953316
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>