uart: uart_stress_tester: Support Zephyr based EC
`chargen` does not currently work on Zephyr based ECs and changes are
going to be made to the fw to fix that.

Get the uart stress test ready for those changes. Specifically, handle
differences in the behavior of the Zephyr shell console. These changes
remain backward compatible with legacy fw.

BUG=b:251872865
BRANCH=none
TEST=uart_stress_tester.py -t 10 -d <raw_ec_uart>

Cq-Depend: chromium:3943262
Change-Id: I15aa307f4b2a0e130b29510e8e0b13134411e23f
Signed-off-by: Mark Hasemeyer <markhas@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3957416
Reviewed-by: Keith Short <keithshort@chromium.org>