host_command: Remove EC_CMD_READ_TEST
The EC_CMD_READ_TEST expected "response_size" to be explicitly set by
the host as greater than zero. This is not how the host commands are
used and there is not a single other host command that behaves in this
way. This expectation resulted in a return value of EC_RES_ERROR every
time the host command was invoked. It's been broken in this way for
about 4 years, which indicates the host command is not used.

Remove the EC_CMD_READ_TEST to gain a bit of flash space back.

BRANCH=none
BUG=b:254534089
TEST=zmake build herobrine image report
TEST=ectool without removal of host command
TEST=CQ

Signed-off-by: Aaron Massey <aaronmassey@google.com>
Change-Id: I2a6cfd4352fb9da1890463a7fadfd0e08c1fb989
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3967307
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-by: Keith Short <keithshort@chromium.org>