TCPMv2: Add host command to send VDM REQ messages
Add an option for the AP to send us VDM REQ messages to pass on to the
port partner or cable.  This is the first of a series of commits, and
later ones will add the transmission of replies to the AP as well as the
underlying EC actions that still need to take place (ex. storing active
modes and such).

LOW_COVERAGE_REASON=Zoss incorrectly marks variable declarations as
uncovered
BRANCH=None
BUG=b:208884535
TEST=./twister -T ./zephyr/test, load on Skyrim and verify messages are
sent as directed to partner and cable with TotalPhase

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I205eaa73e7b4b7aa64c1168ef1d8505f1781a1cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/3969859
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>