zephyr: test: Test `crash` console command
Test various `crash` subcommands that cause different types of fatal
errors. Not all were possible to test due to limitations with running
these on a PC. Also had to make a tweak to the busy loops in the
watchdog and hang subcommands so that they don't interfere with the
simulated sys clock in native_posix. Add a public wrapper function for
calling the crash command outside of the shell interface so we don't
mess up its internal state when the command doesn't return.

BUG=None
BRANCH=None
TEST=./twister

Signed-off-by: Tristan Honscheid <honscheid@google.com>
Change-Id: I393e26441d44c0a5edaf1359413efe6280703add
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4053360
Reviewed-by: Keith Short <keithshort@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>