zephyr: config: select PWM automatically
Add the missing dependency to the PWM subsystem for the drivers that
depends on PWM. This is done using the select Kconfig directive, so that
the subsystem gets enabled explicitly as soon as the first of these gets
defined in the devicetree.

Drop a bunch of now unnecessary CONFIG_PWM=y settings.

BRANCH=none
BUG=none
TEST=cq dry run

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: I55a78d6324f753fe5fe13a0b81b5e129626621af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4031191
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>