TCPMv2: Disallow VCONN Swap in CTAttached.SNK
In PE_VCS_Evaluate_Swap, reject a VCONN Swap when the Type-C SM does not
allow it (when the Type-C state is CTAttached.SNK). Give this setting
precedence over the PE logic to allow VCONN Swap if the TCPM is VCONN
Source in the general case.

BUG=b:257341564
TEST=twister -s zephyr/test/drivers/drivers.usbc_ctvpd
BRANCH=none

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: I2042e5ce5c7694d0efd11b6e7f3c4a14d703676d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4043736
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Sam Hurst <shurst@google.com>