dps: improve option documentation
DPS can behave badly (and is likely to do so) if the configuration doesn't
match actual system characteristics. Existing platforms using DPS seems to
all use the same configuration (possibly because they use the same
charger), but the need for correct configuration could be made clearer.
This rewords the kconfig help for DPS enablement to better describe what
it will do, and what is necessary for a platform to make it work well.

BUG=b:259859051
TEST=zmake build -a
BRANCH=none

Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Change-Id: I481be89d8a00b90e15d01a5c7c8eedbbf057589e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4044793
Reviewed-by: Edward Hill <ecgh@chromium.org>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>