nissa: Configure mux resets for ANX7483
The ANX7483 is only turned on in S5 state and higher, and will default
to a high power consumption state.
Following the skyrim change, set the flag to indicate
that the mux resets in G3 to ensure the USB mux code resets
it to None.

BUG=none
TEST=Check on nivviks that the control register is set to the correct
value (same as at startup) after the AP goes to G3 and is then rebooted.
BRANCH=none

Signed-off-by: Andrew McRae <amcrae@google.com>
Change-Id: I011bcb32028de85153e4c738da999e40624a22ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4050001
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>