zephyr: usb: set a bunch of USBC options automatically
Change most of the USBC options with an associated compatible string to
enable automatically when the corresponding device is defined in the
devicetree.

Most of these don't use the actual struct device yet and only enable a
corresponding ECOS option, and few boards have not been ported to have
the whole usb-c configuration in the devicetree, so some option is a
"default y if ..." rather than a "depends on" so that that can still be
enabled without the device tree node. That can be fixed in the future as
all the boards are ported.

This disables an option for: hoglin, villager, zoglin, zombie due to a
missing dt device (CONFIG_PLATFORM_EC_USBC_PPC_SN5S330) which was
otherwise enabled for all variations.

BRANCH=none
BUG=none
TEST=cq dry run
TEST=zmake compare-builds -a # catches the one above

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>
Change-Id: Ifd543965c2867b35954f48a3fb7afa5289b6e940
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4054626
Reviewed-by: Yuval Peress <peress@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>