rex: fan: Add fan_init method to check if fan is present from CBI
fan_init method is reading FW_FAN value from CBI and deciding if
fan in present for the board. The FW_FAN value is starting at
bit 0 and of size 1. FW_FAN start and size values will be updated
in future when CBI is used in Rex to make decisions as per
information it need to support.

BUG=b:244870433
BRANCH=none
TEST=zmake build rex --clobber &&
./twister -T zephyr/test/rex -i --coverage --clobber

Change-Id: I6aa325453dca497d85099617ab20a3ddb9a08ace
Signed-off-by: Madhurima Paruchuri <mparuchuri@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4111889
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>