ec_commands: Fix EC_CMD_GET_BOOT_TIME response naming
Rename the EC_CMD_GET_BOOT_TIME response struct to 'struct
ec_response_get_boot_time' so it is consistent with the
prevailing naming convention for host commands.

The previous 'struct ap_boot_time_data' name has not propagated to any
other repo (including upstream coreboot), so it is sufficient to rename
it in the EC repo.

BRANCH=none
BUG=b:249163956
TEST='./twister -v -T zephyr/test' passes
	'make buildall' passes
	'make BOARD=host utils-host' passes

Change-Id: Id0c02040b27b66b8e5c8bd1be08f0d59af76468c
Signed-off-by: Caveh Jalali <caveh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4179360
Reviewed-by: Boris Mittelberg <bmbm@google.com>
Reviewed-by: RAJESH KUMAR <rajesh3.kumar@intel.com>