zephyr/shim/led_driver: Add EC_LED_COLOR_INVALID enum value
The ARM EABI uses `-fshort-enums` by default. This causes `enum
ec_led_colors` to be represented as a `uint8_t`. This can cause issues
with the off LED state which is supposed to use -1 for its color.
However, its color ends up as 255 instead. This can cause buffer
overflows in the LED code. Add `EC_LED_COLOR_INVALID` to cause the enum
to be represented as a signed integer and update existing code to use
this value instead of a plain -1.

BRANCH=none
BUG=b:260547985
TEST=Successfully ran LED factory test

Change-Id: Ibc5fd1d46ad408b6e94dc1c3a9b01f269ac8ba7e
Signed-off-by: Robert Zieba <robertzieba@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4179359
Tested-by: Elthan Huang <elthan_huang@compal.corp-partner.google.com>
Reviewed-by: Diana Z <dzigterman@chromium.org>