common: Allow boards to opt out of image padding
The HyperDebug board uses the DFU bootloader in the same way as Servo V4
is going to, meaning that the RO section is 4kB and the RW section takes
up the rest of the 512kB flash space.  Currently, less than 64kB of
actual code is in the RW.

In order to make image files smaller, and speed up DFU flashing by a
factor of ten, I want the option of not having the ec.bin image padded
to the size of the flash space.  This CL introduces a
CONFIG_IMAGE_PADDING, which defaults to being enabled.

TEST=make BOARD=hyperdebug, then flash the smaller image using DFU
BUG=b:192262089

Change-Id: I12cdd4f7b7a3e48c9c7640bafa2d6f10e4ca24f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4193070
Commit-Queue: Jes Klinke <jbk@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>
Reviewed-by: Wai-Hong Tam <waihong@google.com>