Zephyr: mchp: Init spi_config structure properly used by SPI driver
Zephyr SPI drive transfer API requires passing a struct spi_config
containing SPI frequency, mode flags and SPI word size. This structure
was not initialized properly before being passed to the MCHP Zephyr
SPI driver.
The driver checks the fields in the config structure and returned
an error because the word size field was not set to 8-bits.
MCHP SPI HW transfers data in 8-bit chunks.
The fix is to properly initialized the structure in the shim layer
initialize function.

BUG=b:262823672
BRANCH=none
TEST=test on MTL + MEC1727 MECC VerB board

Change-Id: Ib76dab0947870555add427bbbb9f2cbe06e71bd2
Signed-off-by: martin yan <martin.yan@microchip.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4197836
Reviewed-by: Vijay P Hiremath <vijay.p.hiremath@intel.com>
Reviewed-by: Keith Short <keithshort@chromium.org>