usbc-pd: Log with labels should honor debug level
Even if labels are not enabled, honor the debug level
and do not log if the debug level == 0.

It would be better if label support was decoupled from the
logging level, since it may be useful to selectively enable
some level of debugging even without labels (albeit perhaps not
that useful). Logging always when labels are off and debug level == 0
may not be desired, since in some cases the cost of logging is high.

BUG=b:266879860
TEST=Verify on nivviks that logs do not appear when the debug level is
TEST=fixed at 0.
BRANCH=none

Change-Id: Ia271581cf72df2a3598e941b4d984d1472de0c04
Signed-off-by: Andrew McRae <amcrae@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4199236
Reviewed-by: Diana Z <dzigterman@chromium.org>
Commit-Queue: Diana Z <dzigterman@chromium.org>