chip/stm32/clock-stm32l4: Fixed clock cmd documentation
Looking at the code, it is apparent that the "pll" parameter to the
"clock" console command is valid whether STM32_HSE_CLOCK is defined or
not (though the effect may be slightly different).

This CL updates the documentation to list "pll" as a valid option also
when STM32_HSE_CLOCK is not defined.

BUG=none
TEST=make BOARD=hyperdebug

Change-Id: I7da89c15b44fbc3dad2a5fe129b0e361d7f50714
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4216263
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>