core/cortex-m: Allow boards to override interrupt priorities
A lot of code in common/ and elsewhere use DECLARE_IRQ() to designate an
interrupt handler, and set a priority.

This CL makes it possible for board code to override the priorities set
in common code via DECLARE_IRQ(), in case of special considerations.

BUG=b:266832220
TEST=./util/compare_build.sh -b {discovery,munna}

Change-Id: Ied7afe110301b578178159c51033be7fd86abc1f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4216265
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>