system_boot_time: Fix warm reboot counter update for native power seq
Warm reboot counter is being reset to 0 on hook shutdown and increments
on PLTRST. In coldboot scenario, warm reboot counter is incrementing on
PLTRST and hook_shutdown was executed after PLTRST resetting the counter
to 0 resulting the expected value of counter, but with native power
sequencing, hook shutdown is being executed before PLTRST and counter is
not being reset to 0.
This change handles the PLTRST transition appropriately to identify cold
and warm reboot scenario and updates the warm reboot counter as
expected.

BUG=b:249163956
BRANCH=none
TEST='ectool boottime' AP console command is used to fetch data
     'reboot' on AP console to trigger warm reboot
     'apshutdown' on EC console to trigger cold reboot

Change-Id: I969bd9480bbfade481370a5a35f47cbbc53bc4bc
Signed-off-by: Rajesh Kumar <rajesh3.kumar@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4240334
Reviewed-by: Vijay P Hiremath <vijay.p.hiremath@intel.com>
Reviewed-by: Brandon Breitenstein <brandon.breitenstein@intel.com>
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Reviewed-by: caveh jalali <caveh@chromium.org>