skyrim/usbc_config: Ensure that type-A ports are powered after TCPC reset
The GPIO restore code in `reset_nct38xx_port` can disable vbus supply to
the type-A port if the device transitions to S0 during the function
call. Add code to ensure vbus is supplied to the type-A port if we're
in S0.

BRANCH=none
BUG=b:265709281
TEST=ODM verified on frostflow device

Change-Id: I2f4d58801ebbc7c1e5f350a3917e7e2dcc6b5d61
Signed-off-by: Robert Zieba <robertzieba@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4250037
Tested-by: SamSP Liu <samsp_liu2@compal.corp-partner.google.com>
Commit-Queue: Chao Gui <chaogui@google.com>
Reviewed-by: Diana Z <dzigterman@chromium.org>