Zephyr: Remove default-on for retimer FW update
The retimer FW update code is specifically written to support the Intel
USB4 retimers.  Most boards do not use these retimers, so remove the
default-on for it and instead have the retimer presence select the
update module to be included.

BRANCH=None
BUG=None
TEST=CQ+1; build skyrim and observe almost 1k more free flash; zmake
compare-builds passes for rex and herobrine

Change-Id: I1c89bf18b31520a80c4f63a843142efdd5004dd0
Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4249917
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Keith Short <keithshort@chromium.org>