zephyr: de-clang-format a bunch of enums and inits
Usually clang-format works great, but it gets really confused by
macros that include a comma termination in their definition. In those
situation it tends to add unnecessary indentation and makes it more
confusing to understand that the statements on different lines are
actually doing the same thing.

Don't think there's a good way of configuring the tool to do that
properly, and frankly it's probably fine to indent few weird code
snippets manually.

This fixes a bunch of enums and initialization code that had unusual
indentation by shutting clang off for the block and indenting manually

BRANCH=none
BUG=none
TEST=zmake compare-builds -a

Signed-off-by: Fabio Baltieri <fabio.baltieri@gmail.com>
Change-Id: Ic923b27c2ed9f66a2858bfa12e63838f1538700d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4264115
Tested-by: Fabio Baltieri <fabiobaltieri@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Fabio Baltieri <fabiobaltieri@google.com>