board/hyperdebug: Avoid trampling ALTERNATE mode
The HyperDebug board will routinely be asked to reconfigure GPIO pins as
input/open drain/push pull or tweak other aspects of the signals.  It
supports a "reinit" command to re-initialize all GPIO pins according to
their default setting (input or UART/SPI/I2C alternate mode) as declared
in gpio.inc.

However, the reinit logic was flawed.  GPIO_INT_BOTH is declared on
almost every pin, for the case that the user later instructs HyperDebug
to closely monitor a set of pins.  The problem is that GPIO_INT_BOTH
implicitly adds GPIO_INPUT, and this breaks the pins that are declared
as GPIO_ALTERNATE in gpio.inc, because the code in gpio-f0-l.c will
ignore GPIO_ALTERNATE if conflicting GPIO_INPUT is also given.

With this change, SPI, I2C, and UART pins revert to the default non-GPIO
mode, when the "reinit" command is issued.

BUG=266832220
TEST=Observed opentitantool being able to flash via SPI

Change-Id: Ic496ebbda561eb12dbbb506f382f8715c67b8d18
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4295367
Commit-Queue: Jes Klinke <jbk@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>
Reviewed-by: Jett Rink <jettrink@chromium.org>