zephyr: lazor: fix duplicated GPIO
Both the gpio_switchcap_pg_int_l and gpio_da9313_gpio0 entries in the
named-gpios pointed to the same GPIO pin. This causes the GPIO
initialization to configure this pin twice, but only properties from the
second entry are active on the pin.

Use multiple node-labels on the same GPIO pin so that there is only a
single "gpios" entry created.

BUG=b:271576198
BRANCH=none
TEST=zmake build -a

Change-Id: I63a84319b13007d877b13b76a782ea1a6a51f1e9
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4307835
Reviewed-by: Aaron Massey <aaronmassey@google.com>