bma4xx: emulator and driver tests
This implements an emulator and tests for the accel_bma4xx driver using
that emulator in the drivers.default test binary. One change to the driver
itself is made to fix an overread during interrupt processing (which is not
harmful but is slightly inefficient) and a function is marked
__maybe_unused so the driver builds successfully when not using
interrupts.

BUG=b:256047664
TEST=twister -ciC -s drivers/drivers.default -- --test="bma4xx::*"
BRANCH=none

LOW_COVERAGE_REASON=Some aspects of emulator functionality are not
                    currently exercised.
Change-Id: I2765f768c1f5f3651893dfedd1c04cd15cb7fa86
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4310762
Reviewed-by: Eric Yilun Lin <yllin@google.com>