[fix] tbt lost when resume from S4 and S5
In PD test 3.4.4.2,
it will plug tbt devices which is pd-source into unit,
and remove and plug when systen in S4/S5.
Check device list or not when S0.

In this case, if we read port role, cypd will reply as sink mode.
We won't reconnect sink port, so 0xC2 reconnect will fail and cause device lost.

verification: GRL can pass 3.4.4.2 with this CL.

Signed-off-by: elmo_lan <hplan1112@gmail.com>