Lotus: implement the battery recognition and precharge algorithm
When EC	power up, start to pre-charge the battery.
If EC can't detect the battery is present via GPIO or SMBUS after the 30s, stop
to pre-charge the battery.

Due to the ADC table changed, most of the mainboard does not match
the new table, so comment the board id condition in the battery detect function

Signed-off-by: Josh-Tsai <josh_tsai@compal.com>