mt8186,mt8188: fix warm reset issued by servo and gsc
Fix the EC holding SYS_RST_ODL (AP reset) in S3S5. Only asserting the
SYS_RST_ODL with forcing shutdown.  This will allow the AP to boot when
the servo or gsc releases SYS_RST_ODL.

When the AP is in reset, the power_state will be set to S5 first,
and the target should be G3 if the power state has been put to G3 by
S5 inactive timer. This is to prevent EN_PP4200_S5 and PMIC from being
turned off, which would prevent the AP from booting when SYS_RST_ODL
released if the AP reset was held by gsc or servo.

To make it clearer, only three cases that we need to consider:
1. SYS_RST_ODL is held by GSC or Servo
2. SYS_RST_ODL is asserted by AP itself, a gracefully shutdown
3. SYS_RST_ODL is asserted by EC.

For case 1, and 2, we don't need to assert SYS_RST_ODL, since the
EC is not the initiator.

BUG=b:267268982 b:273657518
TEST=On Steelix, and Geralt:
     * Cold reset:
       $ dut-control cold_reset:on sleep:0.2 cold_reset:off
       Result: G3 -> S0
     * Long power press to shutdown:
       $ dut-control dut-control power_key:8.2
       Result: S0 -> S5 -> G3
     * Long power press to power-on but then shutdown:
       $ dut-control dut-control power_key:9.2
       Result: G3 -> S0 -> S5 -> G3
     * Short power press to power-on:
       $ dut-control dut-control power_key:tab
       Result: G3 -> S0
     * Console command: apreset
       Result: S0 -> S0, AP reboots
     * Console command: apshutdown
       Result: S0 -> S5 -> G3
     * Lid open to power-on:
       $ dut-control lid_open:no sleep:0.2 lid_open:yes
       Result: G3 -> S0
     * AP console: reboot
       Reulst: S0 -> S0
     * AP console: poweroff
       Reulst: S0 -> G3
     * Short power press to power-on:
       $ dut-control dut-control power_key:tab
       Result: G3 -> S0
     * Servo issue warm reset:
       $ dut-control power_state:warm_reset
       Result: S0 -> S5 -> S0
     * Servo hold warm_reset and release:
       $ dut-control warm_reset:on sleep:2 warm_reset:off
       Result: S0 -> S5 -> S0
BRANCH=none

Change-Id: I81fc25a5088722487fbbf74d641a5edf4ad450e5
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4338758
Tested-by: Eric Yilun Lin <yllin@google.com>
Commit-Queue: Ting Shen <phoenixshen@chromium.org>
Reviewed-by: Ting Shen <phoenixshen@chromium.org>
Auto-Submit: Eric Yilun Lin <yllin@google.com>