Host sleep: Consider a reboot after suspend to be a resume
When we try to wake the host during a failed resume, the host may reboot
rather than actually following the resume path.  In this case, we will
receive events of "0" indicating we're on a fresh boot and the chipset
RESUME hooks never run.  Instead, we should treat this new boot as a
resume in order to ensure the previous SUSPEND hooks get a RESUME call
to go with them.

BRANCH=None
BUG=b:273327518
TEST=on whiterun, run suspend with bad AP FW version and ensure the
backlight turns on after we wake the host

Change-Id: I9c8e7ad70dbca5245844a31772e99097256e592f
Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4344029
Reviewed-by: Wai-Hong Tam <waihong@google.com>
Commit-Queue: Wai-Hong Tam <waihong@google.com>