chip/stm32: Add extensions for flash programming
Add support to usb_spi.c for processing a new "Flash Command Start"
packet, as an alternative to the existing "Command Start" packet.  The
new serial flash version carries additional flags, among others, to
request a one-byte "Write enable" transaction to take place before the
main transaction, and to request repeated polling of the "busy bit"
after the main transaction.  These allow the number of USB round trips
to be reduced when using Servo Micro or C2D2 for flash programming.

Also, flags of the new request allow the caller to ask for dual or quad
lane communication, which can be supported by board-specific SPI
drivers.  This is intended for use on the HyperDebug debugger board.

BUG=b:273601311
TEST=c2d2 flashing gets slight performance boost

Change-Id: Id4d1ec9f37387b4e083a397ca15300455f7a0ada
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4346952
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>
Reviewed-by: Edward O'Callaghan <quasisec@chromium.org>