include/usb_descriptor.h: Replace wchar_t with uint16_t
We're compiling with -fshort-wchar, which forces wchar_t to be 16-bits:
https://gcc.gnu.org/onlinedocs/gcc/Code-Gen-Options.html.

Ultimately, it looks like the goal was to force wchar_t to be compatible
with the native character type on Windows (UTF-16LE).

Use uint16_t to be explicit about the size we're using, which also
allows us to remove the -fshort-wchar flag.

-fshort-wchar causes issues when linking with libunwind, since libunwind
is not compiled with -fshort-wchar:

ld.lld: error: linking module flags 'wchar_size': IDs have conflicting
values in '/usr/armv7m-cros-eabi/usr/lib/libunwind.a(Unwind-EHABI.cpp.o
at 34472)' and 'ld-temp.o'

This change has no effect on the generated code (see TEST line).

BRANCH=none
BUG=b:274162810
TEST=./util/compare_build.sh -b all -j
     => MATCH
Change-Id: Iab397cb1b83d162d819322aa839e427c4236431b
Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4350955
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>