core/cortex-m: Make barriers in ignore_bus_fault generic
Previously, the barriers were only used on Cortex-M7 because it was
thought to be limited to it:
https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2706039/comment/d33f2fc7_0a027d71/

However, according to
https://developer.arm.com/documentation/ddi0403/d/System-Level-Architecture/System-Level-Programmers--Model/Overview-of-system-level-terminology-and-operation/Exceptions?lang=en,
the Imprecise BusFault is an asynchronous fault in ARMv7-M, so this is
not limited to Cortex-M7.

Without this change the "flash_physical" unit test fails on bloonchipper
(Cortex-M4) when LTO is enabled (see TEST lines).

Also see https://crrev.com/c/1080809 and
https://crrev.com/c/2706039.

BRANCH=none
BUG=b:274162810
TEST=With CONFIG_LTO enabled:
     ./test/run_device_tests.py -b bloonchipper -t flash_physical
     => PASS
Change-Id: Id9da7fbcd9378745439ace80c0b3e48db23f8e4a
Signed-off-by: Tom Hughes <tomhughes@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4356028
Reviewed-by: Diana Z <dzigterman@chromium.org>