zephyr: Add support for rollback regions
Rollback mechanism is used to manage minimal version of RW firmware that
can be booted (to block downgrade attacks when old firmware have
vulnerabilities) and to keep secret (used by fingerprint functionality).

Initial data for rollback regions comes from 'rollback_initial_data'
structure in rollback_initial_data.c. Please note that this file is not
a part of the firmware, but we extract the content of the structure
using objcopy to use it when final image is assembled.

This patch also adds 'rollback0' and 'rollback1' nodes to 'binman.dtsi'
file. This is necessary because rollback regions are usually placed
between WP_RO and EC_RW sections and it's not possible to add nodes
between them after 'binman' node was defined (we can only add nodes at
the end). Binman expects that nodes are ordered by offset, it will
return an error if it's not satisfied.

BUG=b:277029648
TEST=Flash firmware with rollback support. Check if 'rollbackinfo'
     returns the following:

     rollback minimum version: 0
     RW rollback version: 0
     rollback 0: 00000000 00000000 0b112233 [00..00] *
     rollback 1: 00000000 00000000 0b112233 [00..00]

TEST=Run 'rollbackaddent 123' which adds entropy from "123" string to
     rollback region. Check if 'rollbackinfo' returns the following:

     rollback minimum version: 0
     RW rollback version: 0
     rollback 0: 00000000 00000000 0b112233 [00..00]
     rollback 1: 00000001 00000000 0b112233 [f5..a8] *

TEST=Run 'rollbackupdate 2' to set minimum version to 2.
     Check if 'rollbackinfo' returns the following:

     rollback minimum version: 2
     RW rollback version: 0
     rollback 0: 00000002 00000002 0b112233 [f5..a8] *
     rollback 1: 00000001 00000000 0b112233 [f5..a8]

TEST=./twister -v -i -T zephyr/test/rollback --test rollback.default

Change-Id: I44ddf45b84c7cd760489547ad351a626daefa742
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4383044
Commit-Queue: Patryk Duda <patrykd@google.com>
Tested-by: Patryk Duda <patrykd@google.com>
Code-Coverage: Zoss <zoss-cl-coverage@prod.google.com>
Reviewed-by: Eric Yilun Lin <yllin@google.com>