zephyr: nissa: Cleanup leakage GPIOs
SettThe GPIO flags for the power leakage pins to match the flags setup
by the named-gpios node.

The mismatch doesn't currently cause an issue as the GPIO flags are not
used by the Nuvoton power leakage code. This prevents future regressions
if the power leakage code starts using the GPIO flags.

BUG=none
BRANCH=none
TEST=zmake compare-builds -a

Change-Id: I6127d52fe9f583897b325dfa5e64f6fcb5895aa4
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4361321
Reviewed-by: Peter Marheine <pmarheine@chromium.org>