zephyr: uldren: Cleanup leakage GPIOs
Set the GPIO flags for the power leakage pins to match the flags setup
by the named-gpios node.

The mismatch doesn't currently cause an issue as the GPIO flags are not
used by the Nuvoton power leakage code. This prevents future regressions
if the power leakage code starts using the GPIO flags.

BUG=none
BRANCH=none
TEST=zmake compare-builds -a

Change-Id: I019de50f9faf9e30cbcbe5967a6d36602748b7bc
Signed-off-by: Keith Short <keithshort@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4387256
Reviewed-by: Wai-Hong Tam <waihong@google.com>