nissa: implement tests for project common.c
This file contains some utility functions that are fairly straightforward
to test, but requires more complex build configuration than the existing
sub_board tests because it requires the charger code be built in.

The SM5803 input voltage checks are not fully tested because the
isl923x driver does not implement get_battery_cells(); testing that will
require a SM5803 emulator.

BUG=b:267959470
TEST=./twister -T zephyr/test/nissa
BRANCH=nissa

Change-Id: Ic89f8f84b1889cfb4fe588fdd7c6f06a817e76c1
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4402420
Reviewed-by: Tristan Honscheid <honscheid@google.com>