winterhold: Implement custom fan duty adjustment logic
It's strongly based on the existing logic from zephyr/shim/src/fan.c.
The main difference is that the fan duty is changed less aggressively,
in order to avoid fan oscillation. This is done by implementing
a "grace period" functionality. After the duty is change the controller
will wait for a few ticks, depending on how big the change was.
This is a quick fix, at some point it should be replaced by a PID type
of algorithm.

BUG=b:276577324
BRANCH=none
TEST=Verify that the fan doesn't ramp up and down all the time under
     various scenarios, e.g. watching multiple videos on youtube.

Co-developed-by: Radosław Biernacki <biernacki@google.com>
Change-Id: I893a7802df8f5271fe1a253272fdcc3149bdf178
Signed-off-by: Kornel Dulęba <korneld@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4423153
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>
Reviewed-by: Diana Z <dzigterman@chromium.org>