power/icelake: Wait for enough power before booting the AP
Currently, when dibbi is powered from Type-C only, the AP fails to boot
after an EC reboot. This is because it tries to boot the AP before PD
negotiation is complete, so there is not enough power to boot. Modify
the icelake power sequencing to wait for enough power before trying to
boot the AP. This is similar to how CL:1980104 solved the same issue on
puff.

BRANCH=None
BUG=b:272188485
TEST=reboot EC on dibbi with Type-C power only, AP boots successfully

Change-Id: I815c69c58d1f6c9c759ec17b81a7b40fc96279f8
Signed-off-by: Reka Norman <rekanorman@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4421370
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Peter Marheine <pmarheine@chromium.org>