nissa: implement test for preventing buck-boost on SM5803
With the SM5803 emulator landed, this uses the required functionality to
test the board implementation of pd_is_valid_input_voltage(). The
emulator is modified to explicitly cast dt_flags for the interrupt GPIO,
because not doing so causes a warning (promoted to error) if using EC GPIO
flags that lie outside the 16-bit range supported by gpio_dt_spec.

BUG=b:267959470
TEST=twister -T zephyr/test/nissa
BRANCH=nissa

Change-Id: I7192cce5e7e3f08a148cddbf71d68b901d93dc2d
Signed-off-by: Peter Marheine <pmarheine@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4458507
Reviewed-by: Tristan Honscheid <honscheid@google.com>