Charge_state_v2: Always evaluate battery disconnect
Always evaluate the battery disconnect state when we know the battery is
physically present.

Otherwise, we may miss the battery coming out of disconnect if AC is
removed around the same time the battery becomes responsive.  This
causes us to remain in safe state with an active charge port offering
0mV/0mA.

Note this change adjusts the criteria for battery disconnect, counting
errors in assessing the battery disconnect state as disconnects.  This
should account for the fact that the check now runs before we've
verified battery communication is okay.

BRANCH=skyrim
BUG=b:276837557
TEST=on frostflow, ensure units don't show charging when AC is removed
after coming out of ship mode

Change-Id: Ie52c993005cf389dc2912c6ab9492483b60239a9
Signed-off-by: Diana Z <dzigterman@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4470067
Tested-by: SamSP Liu <samsp_liu2@compal.corp-partner.google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>