pre-upload: Make BRANCH= optional, disable Signed-off-by
Signed-off-by is useless. Gerrit enforces that contributors have
signed a CLA in order to upload, and therefore, it's not necessary for
contributors to acknowledge they've signed a CLA in their commit
message.  This would be consistent with the rest of ChromeOS.

Additionally, BRANCH= is not used by the rest of the OS, and as we
move our EC towards a branching model that does not depend on manual
cherry picks, it's becoming less and less useful.  Most commits these
days have stuff like BRANCH=none, or BRANCH=i don't care stop
bothering me.  We could make BRANCH= optional only for commits which
wish to specify it, but not a requirement to upload.

BUG=b:280345831
TEST=pre-upload passes

Change-Id: I1f2f728b9d6890ce2b21dbf3316afe4354ba3054
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4497661
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Jeremy Bettis <jbettis@chromium.org>
Auto-Submit: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Jack Rosenthal <jrosenth@chromium.org>