charger: Move base-external-power check into a function
Move this into a function to shorten the main loop and avoid an #ifdef
in it.

Use a condition check to call the function so that it is clear that it
relates to the base.

Unfortunately we still need the #ifdef in the function due to the
base_responsive variable.

This makes no functional change.

BUG=b:218332694
TEST=zmake build dev-posix

Change-Id: Ia2c689d5800daff43870dc017defa174f93ea38f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4501287
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>