charger: Avoid an #ifdef with base-update function
Take this function outside the #ifdef so we can call it behind an if()
statement. Rename it with a base_ prefix so it is clear which functions
relate to the base, and could perhaps be moved to another file.

This makes no functional change.

Unfortunately we still need the #ifdef in the function due to some
variables which are declared behind that #ifdef.

BUG=b:218332694
TEST=zmake build dev-posix

Change-Id: I102a5b6f98b731b9a2cb42c070b9bd1f20e25f99
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4501288
Commit-Queue: Simon Glass <sjg@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>