sm5803: restore load-bearing ERROR_UNIMPLEMENTED
The OCPC code seems to use ERROR_UNIMPLEMENTED as a signal that it should
do some work, and SUCCESS as a signal that everything has been done.
Always returning success appears to cause incorrect charging behavior:
restore the unconditional ERROR_UNIMPLEMENTED that was removed in commit
ae68afbb5d58ce7af0be3e04f833900b105021cd and change the test to reflect
this.

A note is added to the definition of this driver field to document that
ERROR_UNIMPLEMENTED is an expected normal result.

BUG=b:280530115
TEST=Battery once again charges on Yaviks port C1 when
     connected to a charger.

Change-Id: I491247148c523123206fefcdbcde62a92ea8caf4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4506012
Reviewed-by: Aseda Aboagye <aaboagye@chromium.org>
Tested-by: Peter Marheine <pmarheine@chromium.org>
Commit-Queue: Peter Marheine <pmarheine@chromium.org>