charger: Simplify the arguments to charge_request()
This function is called in two modes, one where it deals with the
currently requested voltage/current and one where it does not. The
function has access to the necessary state so there is no need to
pass it in. It is better to tell the function which approach to take
explicitly, using a boolean, rather than using zero values for the
first two arguments to indicate this.

Replace the first two arguments with a boolean.

This makes no functional change.

BUG=b:218332694
TEST=zmake build dev-posix
Check code on lux:
  *** 69552 bytes in flash and 1152 bytes in RAM lux RO ****
  *** 69416 bytes in flash and 1152 bytes in RAM lux RW ****

Change-Id: I8389e35b0f61b2c4f03a7e7aac746eeab20ef0de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4509872
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>