charger: Avoid accessing curr.ac in base_check_extpower()
We want to move this function into its own file where it will not have
access to the 'curr' state. Pass in the ac value and use the return
value to indicate whether it should be zeroed.

Pass in prev_ac as well, for the same reason.

This changes the ordering of curr.ac being set to 0, so it happens
after board_base_reset() has been called. From what I can tell, this
does not matter, although there is a lot of code what could be
affected by the base_detect_deferred hook or the call to
base_detect_change().

So this makes no functional change.

BUG=b:218332694
TEST=zmake build dev-posix
Check size on lux:
  *** 69552 bytes in flash and 1152 bytes in RAM lux RO ****
  *** 69448 bytes in flash and 1120 bytes in RAM lux RW ****

Change-Id: Ie2171016b692d233fda4f7bd527e74cbb965702e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4510243
Commit-Queue: Simon Glass <sjg@chromium.org>
Reviewed-by: Tomasz Michalec <tmichalec@google.com>
Tested-by: Simon Glass <sjg@chromium.org>