charger: Move setup of input_voltage to charger_base.c
Add a way to set the base input-voltage in the charger_base.c file. Use
this for the two accesses required in charger_base.c, thus removing the
final two base-related #ifdefs in that file.

With this we can drop the inclusion of the ec_ec_comm_client/server
headers.

This makes no functional change.

BUG=b:218332694
TEST=zmake build dev-posix
Check size on lux:
  *** 69552 bytes in flash and 1152 bytes in RAM lux RO ****
  *** 69460 bytes in flash and 1120 bytes in RAM lux RW ****

Check size on brya:
  *** 33464 bytes in flash and 23136 bytes in RAM brya RO ****
  *** 33428 bytes in flash and 23136 bytes in RAM brya RW ****

Change-Id: I8b747b8d7dbe890ce071027aaeed23b879ae8070
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4510250
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>