charger: Add a way to disable debugging at build time
Add a way to disable some debugging related to the chgstate command. It
is a small, but enough to reclaim 70 bytes for damu.

Use functions to avoid needing #ifdefs in the main body of the code.
Convert the value to a bool while we are here.

BUG=b:218332694
TEST=build on brya and see that the feature is still enabled
$ grep CONFIG_CHARGE_DEBUG build/brya/.config
CONFIG_CHARGE_DEBUG=y

Check damu is the same as before:
  *** 12 bytes in flash and 10908 bytes in RAM still available on damu RO ****

Change-Id: I07d2453aa74ab53a29275bba76cd2e8003c5dc14
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4518028
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Simon Glass <sjg@chromium.org>
Tested-by: Simon Glass <sjg@chromium.org>