zephyr: test: Expand body detection test coverage
This intends to verify the parameters used by body detection algorithm.
The logic takes two types of input:

1. RMS noise read from motion sensor driver.
2. var_noise_factor, var_threshold, confidence_delta.

For 1. a mock for motionsense driver was created.
The reasoning behind that is that we don't want driver changes to affect
this test.
For 2. the test will either provide the parameters, or expect the
default values to be used.
Note that this means that if the defaults are changes this test will
have to be updated.

BUG=b:262680246
TEST=./twister -i -s drivers/drivers.body_detection

Change-Id: I0be7ef9c6b082e427850835e4f23631101cc4812
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4518114
Tested-by: Kornel Dulęba <korneld@google.com>
Reviewed-by: Tomasz Michalec <tmichalec@google.com>
Reviewed-by: Keith Short <keithshort@chromium.org>
Commit-Queue: Kornel Dulęba <korneld@google.com>