USB Mux: Use polarity in safe mode
Currently, a handful of drivers look for the mux_state to be equal to
USB_PD_MUX_SAFE_MODE when handling safe mode. This check requires that
the polarity information be cleared before sending the mux_state to any
drivers to correctly handle the safe state. But the same mux_state is
sent to the kernel, which relies on polarity information from the EC
when setting USB-C switches.

This CL updates the retimer, tcpm and usb_mux drivers which check for
mux_state to equal USB_PD_MUX_SAFE_MODE to instead only look for the
safe mode bit to be set. This allows those drivers to function the same
way in safe mode regardless of polarity information. It also updates
usb_mux.c to preserve the polarity bit when sending mux_state to drivers
in safe_state.

LOW_COVERAGE_REASON=20 of the changed drivers have no emulators
BUG=b:279733113
TEST=make -j buildall, confirmed that no remaining drivers check for mux
state to equal USB_PD_MUX_SAFE_MODE. Checked on redrix that we are able
to drive a monitor with inverted and normal polarity after losing
Try.SRC.

Change-Id: Ifc89d0ba1da0975436f93be102fadf725a5b8566
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4527219
Reviewed-by: Diana Z <dzigterman@chromium.org>
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
Tested-by: Jameson Thies <jthies@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>