npcx: cec: Fix bug in receiving packets not destined to us
Currently, when receiving a packet not destined to us, there is no
timeout set in CEC_STATE_FOLLOWER_ACK_LOW. The reason this hasn't caused
problems is that the timer stays in whatever state it was in previously,
which in practice happens to be with the timer still running from the
previous state. So we get a timeout at some point in the future,
consider it an error and return to the IDLE state. And since the packet
is not destined to us, it doesn't matter that we didn't receive it
successfully.

However, we shouldn't rely on the previous timer state, and it's
possible for the state machine to get stuck since there's no timeout
(and this happened while working on the ITE version). So instead set a
timeout to keep reading packets not destined to us. The comment "Don't
ack broadcast or packets which destination are us, but continue reading"
suggests that this was always the intended behaviour.

BUG=b:270507438
TEST=Add logging to check that packets not destined to us are now
received successfully on kinox (but are still not passed to the host due
to the check in CEC_STATE_FOLLOWER_ACK_FINISH).

Change-Id: I0a06198e121f5e57e683b6ac1825c8cf58e6d1ca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4598631
Reviewed-by: Stefan Adolfsson <sadolfsson@chromium.org>
Reviewed-by: Daisuke Nojiri <dnojiri@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>