board/hyperdebug: Add declarations of I2C4
HyperDebug supports four I2C busses, three of which are to be connected
to the three I2C busses used by Ti50.  The fourth we want to use for
connecting a power measurement chip on the shield, in order to
automatically verify that the power draw in sleep modes does not have
regressions.

BUG=b:282999020
TEST=Observe HyperDebug manipulate IO expanders via I2C4

Change-Id: Ibc4eeab692e0278eb3cc0c77c1affd035b555bc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4537012
Reviewed-by: Jett Rink <jettrink@chromium.org>
Tested-by: Jes Klinke <jbk@chromium.org>
Commit-Queue: Jes Klinke <jbk@chromium.org>