starmie: Add AC/DC feedback lcd pin
Because Himax IC needs to set different touch
parameters according to the charging state, so
here set up a HOOK_AC_CHANGE event. when AC is
plugged in, EC_GPM6 is set to high level and
when AC power is removed,EC_GPM6 is set to low
level.

BUG=b:282117960
TEST=zmake build starmie
The test results are consistent with the expected
behavior.

Change-Id: Ief963661540e23d67147a17f4ecee464824e7000
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4542365
Reviewed-by: Ruihai Zhou <zhouruihai@huaqin.corp-partner.google.com>
Reviewed-by: Sung-Chi Li <lschyi@chromium.org>
Tested-by: Ganxiang Wang <wangganxiang@huaqin.corp-partner.google.com>
Commit-Queue: Ganxiang Wang <wangganxiang@huaqin.corp-partner.google.com>