rex: Declare TCPC, PPC, MUX on USB3 DB alt
Rex has multiple USB daughterboard options. It doesn't really make sense
to have a default TCPC, PPC, or MUX for one DB and use "is-alt" for the
others. So, set up all TCPCs and PPCs as "is-alt" and select the
appropriate one at runtime. Similarly for the DB MUX, except that much
chains are declared as "alternate-chain" instead of "is-alt".

BUG=b:267589042
TEST=boots on rex
TEST='./twister -v -T zephyr/test/rex' passes

Change-Id: I7bb524c66825c0dd1c36ae8137b260e5b78234ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/4550239
Reviewed-by: Keith Short <keithshort@chromium.org>
Tested-by: caveh jalali <caveh@chromium.org>
Commit-Queue: caveh jalali <caveh@chromium.org>
Auto-Submit: caveh jalali <caveh@chromium.org>
Reviewed-by: Fabio Baltieri <fabiobaltieri@google.com>